--- Tests/TMDBTests/CoverageTests.swift
@@ -1,18 +1,0 @@
-import TMDB
-import Testing
-
-struct CoverageTests {
-  @Test func completeMissing() {
-    // This is only meant to ensure all code is covered
-    _ = Country()
-    _ = Genre()
-    _ = Keyword()
-    _ = Language()
-    _ = MovieCollection()
-    _ = MovieDetails()
-    _ = MovieDetails.Keywords()
-    _ = ProductionCompany()
-    _ = TVShowDetails()
-    _ = TVShowDetails.Keywords()
-  }
-}
