--- Tests/TMDBTests/Movies/Data+Movie1.swift
@@ -1,24 +1,0 @@
-import Foundation
-
-extension Data {
-  static let movie1 = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "adult": false,
-    "backdrop_path": null,
-    "genre_ids": [],
-    "id": 1524520,
-    "original_language": "cs",
-    "original_title": "Conversation 5, 6 & 8",
-    "overview": "",
-    "popularity": 0,
-    "poster_path": null,
-    "release_date": "1985-01-01",
-    "title": "Conversation 5, 6 & 8",
-    "video": false,
-    "vote_average": 0,
-    "vote_count": 0
-  }
-  """#
