--- Tests/TMDBTests/Movies/Data+Movie2.swift
@@ -1,18 +1,0 @@
-import Foundation
-
-extension Data {
-  static let movie2 = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "adult": false,
-    "backdrop_path": null,
-    "id": 1524521,
-    "original_language": "nl",
-    "original_title": "Scrooge live  collectie",
-    "overview": "",
-    "poster_path": "/jhautnzIZYPzNgNy8ZIQRGaZrun.jpg",
-    "title": "Scrooge live  collectie"
-  }
-  """#
