--- Tests/TMDBTests/Movies/Data+MovieDetailsComplete.swift
@@ -1,532 +1,0 @@
-import Foundation
-
-extension Data {
-  static let movieDetailsComplete = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "adult": false,
-    "backdrop_path": "/1RAxtBxslR4OZCZC1vxIRUxjR7a.jpg",
-    "belongs_to_collection": {
-      "id": 86860,
-      "name": "Scooby-Doo Collection",
-      "poster_path": "/GoZ2F80wVsnvdXtwtWY5HPnxYG.jpg",
-      "backdrop_path": "/crBcWcmfSwM1FaFr7s3dOyzSF9m.jpg"
-    },
-    "budget": 84000000,
-    "genres": [
-      {
-        "id": 9648,
-        "name": "Mystery"
-      },
-      {
-        "id": 12,
-        "name": "Adventure"
-      },
-      {
-        "id": 35,
-        "name": "Comedy"
-      }
-    ],
-    "homepage": "https://www.warnerbros.com/movies/scooby-doo-movie",
-    "id": 9637,
-    "imdb_id": "tt0267913",
-    "origin_country": [
-      "US"
-    ],
-    "original_language": "en",
-    "original_title": "Scooby-Doo",
-    "overview": "When the Mystery Inc. gang is invited to Spooky Island, a popular amusement park, they soon discover that the attractions aren't the only things that are spooky. Strange things are happening, and it's up to Scooby, Shaggy, Fred, Daphne, and Velma to uncover the truth behind the mysterious happenings.",
-    "popularity": 6.5838,
-    "poster_path": "/mTAiBJGg8mqEfnYHHbi37ZoRSZm.jpg",
-    "production_companies": [
-      {
-        "id": 748,
-        "logo_path": null,
-        "name": "Mosaic Media Group",
-        "origin_country": "US"
-      },
-      {
-        "id": 174,
-        "logo_path": "/kgJaIcKZZK4lTazer55XrgWDEvp.png",
-        "name": "Warner Bros. Pictures",
-        "origin_country": "US"
-      }
-    ],
-    "production_countries": [
-      {
-        "iso_3166_1": "US",
-        "name": "United States of America"
-      }
-    ],
-    "release_date": "2002-06-14",
-    "revenue": 275650703,
-    "runtime": 87,
-    "spoken_languages": [
-      {
-        "english_name": "English",
-        "iso_639_1": "en",
-        "name": "English"
-      }
-    ],
-    "status": "Released",
-    "tagline": "Be afraid. Be kind of afraid.",
-    "title": "Scooby-Doo",
-    "video": false,
-    "vote_average": 6.092,
-    "vote_count": 4562,
-    "keywords": {
-      "keywords": [
-        {
-          "id": 3392,
-          "name": "voodoo"
-        },
-        {
-          "id": 703,
-          "name": "detective"
-        },
-        {
-          "id": 11478,
-          "name": "talking dog"
-        },
-        {
-          "id": 10244,
-          "name": "based on cartoon"
-        },
-        {
-          "id": 10793,
-          "name": "resort"
-        },
-        {
-          "id": 157241,
-          "name": "criminal investigation"
-        },
-        {
-          "id": 158340,
-          "name": "theme park"
-        },
-        {
-          "id": 164246,
-          "name": "nostalgic"
-        },
-        {
-          "id": 179431,
-          "name": "duringcreditsstinger"
-        },
-        {
-          "id": 209220,
-          "name": "live action and animation"
-        }
-      ]
-    },
-    "similar": {
-      "page": 1,
-      "results": [
-        {
-          "adult": false,
-          "backdrop_path": "/uif5fUshJrXyyDzfpzp1DLw3N0S.jpg",
-          "genre_ids": [
-            27,
-            53,
-            9648
-          ],
-          "id": 539,
-          "original_language": "en",
-          "original_title": "Psycho",
-          "overview": "When larcenous real estate clerk Marion Crane goes on the lam with a wad of cash and hopes of starting a new life, she ends up at the notorious Bates Motel, where manager Norman Bates cares for his housebound mother.",
-          "popularity": 9.4094,
-          "poster_path": "/yz4QVqPx3h1hD1DfqqQkCq3rmxW.jpg",
-          "release_date": "1960-06-22",
-          "title": "Psycho",
-          "video": false,
-          "vote_average": 8.421,
-          "vote_count": 10450
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/c7VbRPlev1bRfKvTxr98erw5T7B.jpg",
-          "genre_ids": [
-            16,
-            35,
-            10751
-          ],
-          "id": 585,
-          "original_language": "en",
-          "original_title": "Monsters, Inc.",
-          "overview": "Lovable Sulley and his wisecracking sidekick Mike Wazowski are the top scare team at Monsters, Inc., the scream-processing factory in Monstropolis. When a little girl named Boo wanders into their world, it's the monsters who are scared silly, and it's up to Sulley and Mike to keep her out of sight and get her back home.",
-          "popularity": 16.9668,
-          "poster_path": "/wFSpyMsp7H0ttERbxY7Trlv8xry.jpg",
-          "release_date": "2001-11-01",
-          "title": "Monsters, Inc.",
-          "video": false,
-          "vote_average": 7.845,
-          "vote_count": 19001
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/qVKhr3WcbSQKJmVJhKq9ov3iwcT.jpg",
-          "genre_ids": [
-            9648,
-            53,
-            18
-          ],
-          "id": 617,
-          "original_language": "en",
-          "original_title": "Wild Things",
-          "overview": "When teen-socialite Kelly Van Ryan and troubled bad girl Suzie Toller accuse guidance counselor Sam Lombardo of rape, he's suspended by the school, rejected by the town, and fighting to get his life back. One cop suspects conspiracy, but nothing is what it seems...",
-          "popularity": 3.4378,
-          "poster_path": "/wrcTDD9T7Ga5c9MW7kaOo2qwIvW.jpg",
-          "release_date": "1998-03-20",
-          "title": "Wild Things",
-          "video": false,
-          "vote_average": 6.358,
-          "vote_count": 1829
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/c6yfABGVKuB5cjoOwdX4AJMlzUz.jpg",
-          "genre_ids": [
-            35,
-            14
-          ],
-          "id": 620,
-          "original_language": "en",
-          "original_title": "Ghostbusters",
-          "overview": "After losing their academic posts at a prestigious university, a team of parapsychologists goes into business as proton-pack-toting \"ghostbusters\" who exterminate ghouls, hobgoblins and supernatural pests of all stripes. An ad campaign pays off when a knockout cellist hires the squad to purge her swanky digs of demons that appear to be living in her refrigerator.",
-          "popularity": 7.5124,
-          "poster_path": "/7E8nLijS9AwwUEPu2oFYOVKhdFA.jpg",
-          "release_date": "1984-06-08",
-          "title": "Ghostbusters",
-          "video": false,
-          "vote_average": 7.494,
-          "vote_count": 9301
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/imvV078HyWsa9jgDVLi4APgskrJ.jpg",
-          "genre_ids": [
-            35,
-            10749,
-            18
-          ],
-          "id": 634,
-          "original_language": "en",
-          "original_title": "Bridget Jones's Diary",
-          "overview": "Bridget Jones is an average woman struggling against her age, her weight, her job, her lack of a man, and her various imperfections. As a New Year's resolution, Bridget decides to take control of her life, starting by keeping a diary in which she will always tell the complete truth. The fireworks begin when her charming though disreputable boss takes an interest in the quirky Miss Jones. Thrown into the mix are Bridget's band of slightly eccentric friends and a rather disagreeable acquaintance into whom Bridget cannot seem to stop running or help finding quietly attractive.",
-          "popularity": 4.0946,
-          "poster_path": "/dkauRl9TosBFikftrC3OVcKWDoz.jpg",
-          "release_date": "2001-04-13",
-          "title": "Bridget Jones's Diary",
-          "video": false,
-          "vote_average": 6.747,
-          "vote_count": 5261
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/viDonz5uLmengw9RUAh3Rf65dQG.jpg",
-          "genre_ids": [
-            27,
-            9648
-          ],
-          "id": 635,
-          "original_language": "en",
-          "original_title": "Angel Heart",
-          "overview": "A down-and-out Brooklyn detective is hired to track down a singer on an odyssey that will take him through the desperate streets of Harlem, the smoke-filled jazz clubs of New Orleans, and the swamps of Louisiana and its seedy underworld of voodoo.",
-          "popularity": 2.1523,
-          "poster_path": "/h5v3wjJQNB7q2RntEnKDLhKtTFE.jpg",
-          "release_date": "1987-03-06",
-          "title": "Angel Heart",
-          "video": false,
-          "vote_average": 7.127,
-          "vote_count": 1612
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/y8EijsAJLyipw9cbbcMmdkPEGK8.jpg",
-          "genre_ids": [
-            16,
-            10751,
-            28,
-            12,
-            35,
-            878
-          ],
-          "id": 616473,
-          "original_language": "ms",
-          "original_title": "BoBoiBoy Movie 2",
-          "overview": "BoBoiBoy and his friends must protect his elemental powers from an ancient villain seeking to regain control and wreak cosmic havoc.",
-          "popularity": 0.2628,
-          "poster_path": "/5wFIkx7aLhTHxwztj8AXgF17rHj.jpg",
-          "release_date": "2019-08-02",
-          "title": "BoBoiBoy: Elemental Heroes",
-          "video": true,
-          "vote_average": 8.5,
-          "vote_count": 15
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/8uNjWHsSW3TSOWDU4EY76H5MzAG.jpg",
-          "genre_ids": [
-            28,
-            53,
-            12
-          ],
-          "id": 657,
-          "original_language": "en",
-          "original_title": "From Russia with Love",
-          "overview": "Agent 007 is back in the second installment of the James Bond series, this time battling a secret crime organization known as SPECTRE. Russians Rosa Klebb and Kronsteen are out to snatch a decoding device known as the Lektor, using the ravishing Tatiana to lure Bond into helping them. Bond willingly travels to meet Tatiana in Istanbul, where he must rely on his wits to escape with his life in a series of deadly encounters with the enemy.",
-          "popularity": 5.8012,
-          "poster_path": "/v8YXfFwfGMpYLlDQdBQ1jNhehmY.jpg",
-          "release_date": "1963-10-10",
-          "title": "From Russia with Love",
-          "video": false,
-          "vote_average": 7.074,
-          "vote_count": 3130
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/mM4WeXqJd2tu1gCF5GYyAZf2Npl.jpg",
-          "genre_ids": [
-            12,
-            28,
-            53
-          ],
-          "id": 658,
-          "original_language": "en",
-          "original_title": "Goldfinger",
-          "overview": "Special agent 007 comes face to face with one of the most notorious villains of all time, and now he must outwit and outgun the powerful tycoon to prevent him from cashing in on a devious scheme to raid Fort Knox -- and obliterate the world's economy.",
-          "popularity": 5.6748,
-          "poster_path": "/4CD9eftOlBcutoXak5C7L24X1Y2.jpg",
-          "release_date": "1964-09-20",
-          "title": "Goldfinger",
-          "video": false,
-          "vote_average": 7.355,
-          "vote_count": 3717
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/9qGWOflC5WNhaKk2UPYu55SjjNj.jpg",
-          "genre_ids": [
-            80,
-            53,
-            9648
-          ],
-          "id": 678,
-          "original_language": "en",
-          "original_title": "Out of the Past",
-          "overview": "The peaceful life of a gas station owner is disrupted when a man from his past arrives in town and forces him to return to the dark world he had tried to escape.",
-          "popularity": 1.845,
-          "poster_path": "/lnUK6Cg5ARM0qaq0B5SG20VAM0h.jpg",
-          "release_date": "1947-11-25",
-          "title": "Out of the Past",
-          "video": false,
-          "vote_average": 7.6,
-          "vote_count": 616
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/jzMhC6NB4KHqfSAgbW2BFslLuSh.jpg",
-          "genre_ids": [
-            10751,
-            35,
-            12,
-            16
-          ],
-          "id": 991494,
-          "original_language": "en",
-          "original_title": "The SpongeBob Movie: Search for SquarePants",
-          "overview": "SpongeBob travels to the depths of the ocean and faces off against the ghost of the Flying Dutchman.",
-          "popularity": 5.0398,
-          "poster_path": "/e6AkALFDWytfXslFJ7s2l8Xk4TX.jpg",
-          "release_date": "2025-12-18",
-          "title": "The SpongeBob Movie: Search for SquarePants",
-          "video": false,
-          "vote_average": 0,
-          "vote_count": 0
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/eWMHecmPAbOCqALSrq45Jqj4BhF.jpg",
-          "genre_ids": [
-            10751,
-            35,
-            16,
-            12,
-            10770,
-            14
-          ],
-          "id": 439058,
-          "original_language": "en",
-          "original_title": "Hey Arnold! The Jungle Movie",
-          "overview": "When their trip to San Lorenzo takes a turn for the worst, Arnold and his classmate’s only hope of getting home is retracing the dangerous path that led to Arnold's parents' disappearance.",
-          "popularity": 0.9915,
-          "poster_path": "/asBhWDiqGxelve0DtMfMFIzR0Jm.jpg",
-          "release_date": "2018-04-26",
-          "title": "Hey Arnold! The Jungle Movie",
-          "video": false,
-          "vote_average": 7.628,
-          "vote_count": 188
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/dBSYDUZpps3LWru7ofUZAYoFgUj.jpg",
-          "genre_ids": [
-            9648,
-            53
-          ],
-          "id": 1586,
-          "original_language": "en",
-          "original_title": "Secret Window",
-          "overview": "Mort Rainey, a writer just emerging from a painful divorce with his ex-wife, is stalked at his remote lake house by a psychotic stranger and would-be scribe who claims Rainey swiped his best story idea. But as Rainey endeavors to prove his innocence, he begins to question his own sanity.",
-          "popularity": 2.801,
-          "poster_path": "/hvzw4wufKZkDCTPVeG71Z7pGfZh.jpg",
-          "release_date": "2004-03-12",
-          "title": "Secret Window",
-          "video": false,
-          "vote_average": 6.564,
-          "vote_count": 3754
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/3XpxR3cSQJou79KBQpPCcpZzmhj.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 1588,
-          "original_language": "en",
-          "original_title": "Bring It On",
-          "overview": "The Toro cheerleading squad from Rancho Carne High School in San Diego has got spirit, spunk, sass and a killer routine that's sure to land them the national championship trophy for the sixth year in a row. But for newly-elected team captain Torrance, the Toros' road to total cheer glory takes a shady turn when she discovers that their perfectly-choreographed routines were in fact stolen.",
-          "popularity": 3.1445,
-          "poster_path": "/9SkQeC55tUwtQU0rSplXBBbVfty.jpg",
-          "release_date": "2000-08-25",
-          "title": "Bring It On",
-          "video": false,
-          "vote_average": 6.253,
-          "vote_count": 1753
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/1ZVlzOBYCggLzvVYHK7Im6LohjC.jpg",
-          "genre_ids": [
-            28,
-            12,
-            35,
-            10751,
-            14
-          ],
-          "id": 1593,
-          "original_language": "en",
-          "original_title": "Night at the Museum",
-          "overview": "Chaos reigns at the natural history museum when night watchman Larry Daley accidentally stirs up an ancient curse, awakening Attila the Hun, an army of gladiators, a Tyrannosaurus rex and other exhibits.",
-          "popularity": 7.0164,
-          "poster_path": "/pDsAAYf6Zn0yiAGJ6lYGs6hoZ4E.jpg",
-          "release_date": "2006-12-20",
-          "title": "Night at the Museum",
-          "video": false,
-          "vote_average": 6.59,
-          "vote_count": 10494
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/lMSlZxJp43HR81XacWh75w3hW2l.jpg",
-          "genre_ids": [
-            35,
-            10749
-          ],
-          "id": 1599,
-          "original_language": "en",
-          "original_title": "I ♥ Huckabees",
-          "overview": "A husband-and-wife team play detective, but not in the traditional sense. Instead, the happy duo helps others solve their existential issues, the kind that keep you up at night, wondering what it all means.",
-          "popularity": 1.3561,
-          "poster_path": "/h27ytqXyOrAB6KmRx2abOp8p2W.jpg",
-          "release_date": "2004-09-10",
-          "title": "I ♥ Huckabees",
-          "video": false,
-          "vote_average": 6.282,
-          "vote_count": 666
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/7FbUtL04mRyMAUkcXjNNz8w6rbf.jpg",
-          "genre_ids": [
-            35,
-            10751
-          ],
-          "id": 263387,
-          "original_language": "en",
-          "original_title": "Ernest Goes to Splash Mountain",
-          "overview": "An exciting mock documentary produced for the opening of the Splash Mountain attraction at Disneyland. Ernest Goes To Splash Mountain chronicles the adventures of Ernest P. Worrell who trains as America's first \"Splashtranaut\" destined to become the very first person to conquer the mountain. Ralph Story, veteran news anchor and his team are determined to cover the momentous event by every angle possible.",
-          "popularity": 0.4058,
-          "poster_path": "/jtoJYUpLZs1erzVjyfo59xCIYDP.jpg",
-          "release_date": "1989-07-07",
-          "title": "Ernest Goes to Splash Mountain",
-          "video": false,
-          "vote_average": 4.6,
-          "vote_count": 11
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/ocCAwBuaRqMuYDstTQihkqg6zsz.jpg",
-          "genre_ids": [
-            28,
-            35,
-            80
-          ],
-          "id": 5175,
-          "original_language": "en",
-          "original_title": "Rush Hour 2",
-          "overview": "It's vacation time for Carter as he finds himself alongside Lee in Hong Kong wishing for more excitement. While Carter wants to party and meet the ladies, Lee is out to track down a Triad gang lord who may be responsible for killing two men at the American Embassy. Things get complicated as the pair stumble onto a counterfeiting plot. The boys are soon up to their necks in fist fights and life-threatening situations. A trip back to the U.S. may provide the answers about the bombing, the counterfeiting, and the true allegiance of sexy customs agent Isabella.",
-          "popularity": 8.0126,
-          "poster_path": "/aBQf2vMiCINeVC9v6BGVYKXurTh.jpg",
-          "release_date": "2001-08-03",
-          "title": "Rush Hour 2",
-          "video": false,
-          "vote_average": 6.752,
-          "vote_count": 4309
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/x29PYzoNYPGzdj3M8FDwSwEgNpf.jpg",
-          "genre_ids": [
-            35,
-            80,
-            9648,
-            53
-          ],
-          "id": 5236,
-          "original_language": "en",
-          "original_title": "Kiss Kiss Bang Bang",
-          "overview": "A petty thief posing as an actor is brought to Los Angeles for an unlikely audition and finds himself in the middle of a murder investigation along with his high school dream girl and a detective who's been training him for his upcoming role...",
-          "popularity": 1.7526,
-          "poster_path": "/aWfjIkpENFX6Uw82pET7EQ6jnrd.jpg",
-          "release_date": "2005-09-14",
-          "title": "Kiss Kiss Bang Bang",
-          "video": false,
-          "vote_average": 7.166,
-          "vote_count": 2812
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/blHK98x8CZoG2v5k5E6b1o2RmvF.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 441940,
-          "original_language": "en",
-          "original_title": "How DOooo You Do",
-          "overview": "Murder occurs when several of the most popular radio personalities of the '40s converge on a desert resort.",
-          "popularity": 0.1396,
-          "poster_path": "/nipPuS3Ac8XWE1WDJIt51ZmKUEQ.jpg",
-          "release_date": "1945-12-24",
-          "title": "How DOooo You Do",
-          "video": false,
-          "vote_average": 0,
-          "vote_count": 0
-        }
-      ],
-      "total_pages": 10052,
-      "total_results": 201033
-    }
-  }
-  """#
