+++ Tests/TMDBTests/Movies/Data+MovieDetailsMock.swift
@@ -1,0 +1,237 @@
+//
+//  Data+MovieDetailsMock.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+import Foundation
+
+extension Data {
+  static let movieDetailsMock = Data(content.utf8)
+}
+
+private let content = #"""
+  {
+    "adult": false,
+    "backdrop_path": "/1RAxtBxslR4OZCZC1vxIRUxjR7a.jpg",
+    "belongs_to_collection": {
+      "id": 86860,
+      "name": "Scooby-Doo Collection",
+      "poster_path": "/GoZ2F80wVsnvdXtwtWY5HPnxYG.jpg",
+      "backdrop_path": "/crBcWcmfSwM1FaFr7s3dOyzSF9m.jpg"
+    },
+    "budget": 84000000,
+    "genres": [
+      {
+        "id": 9648,
+        "name": "Mystery"
+      },
+      {
+        "id": 12,
+        "name": "Adventure"
+      },
+      {
+        "id": 35,
+        "name": "Comedy"
+      }
+    ],
+    "homepage": "https://www.warnerbros.com/movies/scooby-doo-movie",
+    "id": 9637,
+    "imdb_id": "tt0267913",
+    "origin_country": [
+      "US"
+    ],
+    "original_language": "en",
+    "original_title": "Scooby-Doo",
+    "overview": "When the Mystery Inc. gang is invited to Spooky Island, a popular amusement park, they soon discover that the attractions aren't the only things that are spooky. Strange things are happening, and it's up to Scooby, Shaggy, Fred, Daphne, and Velma to uncover the truth behind the mysterious happenings.",
+    "popularity": 6.5838,
+    "poster_path": "/mTAiBJGg8mqEfnYHHbi37ZoRSZm.jpg",
+    "production_companies": [
+      {
+        "id": 748,
+        "logo_path": null,
+        "name": "Mosaic Media Group",
+        "origin_country": "US"
+      },
+      {
+        "id": 174,
+        "logo_path": "/kgJaIcKZZK4lTazer55XrgWDEvp.png",
+        "name": "Warner Bros. Pictures",
+        "origin_country": "US"
+      }
+    ],
+    "production_countries": [
+      {
+        "iso_3166_1": "US",
+        "name": "United States of America"
+      }
+    ],
+    "release_date": "2002-06-14",
+    "revenue": 275650703,
+    "runtime": 87,
+    "spoken_languages": [
+      {
+        "english_name": "English",
+        "iso_639_1": "en",
+        "name": "English"
+      }
+    ],
+    "status": "Released",
+    "tagline": "Be afraid. Be kind of afraid.",
+    "title": "Scooby-Doo",
+    "video": false,
+    "vote_average": 6.092,
+    "vote_count": 4562,
+    "keywords": {
+      "keywords": [
+        {
+          "id": 3392,
+          "name": "voodoo"
+        },
+        {
+          "id": 703,
+          "name": "detective"
+        },
+        {
+          "id": 11478,
+          "name": "talking dog"
+        },
+        {
+          "id": 10244,
+          "name": "based on cartoon"
+        },
+        {
+          "id": 10793,
+          "name": "resort"
+        },
+        {
+          "id": 157241,
+          "name": "criminal investigation"
+        },
+        {
+          "id": 158340,
+          "name": "theme park"
+        },
+        {
+          "id": 164246,
+          "name": "nostalgic"
+        },
+        {
+          "id": 179431,
+          "name": "duringcreditsstinger"
+        },
+        {
+          "id": 209220,
+          "name": "live action and animation"
+        }
+      ]
+    },
+    "similar": {
+      "page": 1,
+      "results": [
+        {
+          "adult": false,
+          "backdrop_path": "/uif5fUshJrXyyDzfpzp1DLw3N0S.jpg",
+          "genre_ids": [
+            27,
+            53,
+            9648
+          ],
+          "id": 539,
+          "original_language": "en",
+          "original_title": "Psycho",
+          "overview": "When larcenous real estate clerk Marion Crane goes on the lam with a wad of cash and hopes of starting a new life, she ends up at the notorious Bates Motel, where manager Norman Bates cares for his housebound mother.",
+          "popularity": 9.4094,
+          "poster_path": "/yz4QVqPx3h1hD1DfqqQkCq3rmxW.jpg",
+          "release_date": "1960-06-22",
+          "title": "Psycho",
+          "video": false,
+          "vote_average": 8.421,
+          "vote_count": 10450
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/c7VbRPlev1bRfKvTxr98erw5T7B.jpg",
+          "genre_ids": [
+            16,
+            35,
+            10751
+          ],
+          "id": 585,
+          "original_language": "en",
+          "original_title": "Monsters, Inc.",
+          "overview": "Lovable Sulley and his wisecracking sidekick Mike Wazowski are the top scare team at Monsters, Inc., the scream-processing factory in Monstropolis. When a little girl named Boo wanders into their world, it's the monsters who are scared silly, and it's up to Sulley and Mike to keep her out of sight and get her back home.",
+          "popularity": 16.9668,
+          "poster_path": "/wFSpyMsp7H0ttERbxY7Trlv8xry.jpg",
+          "release_date": "2001-11-01",
+          "title": "Monsters, Inc.",
+          "video": false,
+          "vote_average": 7.845,
+          "vote_count": 19001
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/qVKhr3WcbSQKJmVJhKq9ov3iwcT.jpg",
+          "genre_ids": [
+            9648,
+            53,
+            18
+          ],
+          "id": 617,
+          "original_language": "en",
+          "original_title": "Wild Things",
+          "overview": "When teen-socialite Kelly Van Ryan and troubled bad girl Suzie Toller accuse guidance counselor Sam Lombardo of rape, he's suspended by the school, rejected by the town, and fighting to get his life back. One cop suspects conspiracy, but nothing is what it seems...",
+          "popularity": 3.4378,
+          "poster_path": "/wrcTDD9T7Ga5c9MW7kaOo2qwIvW.jpg",
+          "release_date": "1998-03-20",
+          "title": "Wild Things",
+          "video": false,
+          "vote_average": 6.358,
+          "vote_count": 1829
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/c6yfABGVKuB5cjoOwdX4AJMlzUz.jpg",
+          "genre_ids": [
+            35,
+            14
+          ],
+          "id": 620,
+          "original_language": "en",
+          "original_title": "Ghostbusters",
+          "overview": "After losing their academic posts at a prestigious university, a team of parapsychologists goes into business as proton-pack-toting \"ghostbusters\" who exterminate ghouls, hobgoblins and supernatural pests of all stripes. An ad campaign pays off when a knockout cellist hires the squad to purge her swanky digs of demons that appear to be living in her refrigerator.",
+          "popularity": 7.5124,
+          "poster_path": "/7E8nLijS9AwwUEPu2oFYOVKhdFA.jpg",
+          "release_date": "1984-06-08",
+          "title": "Ghostbusters",
+          "video": false,
+          "vote_average": 7.494,
+          "vote_count": 9301
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/imvV078HyWsa9jgDVLi4APgskrJ.jpg",
+          "genre_ids": [
+            35,
+            10749,
+            18
+          ],
+          "id": 634,
+          "original_language": "en",
+          "original_title": "Bridget Jones's Diary",
+          "overview": "Bridget Jones is an average woman struggling against her age, her weight, her job, her lack of a man, and her various imperfections. As a New Year's resolution, Bridget decides to take control of her life, starting by keeping a diary in which she will always tell the complete truth. The fireworks begin when her charming though disreputable boss takes an interest in the quirky Miss Jones. Thrown into the mix are Bridget's band of slightly eccentric friends and a rather disagreeable acquaintance into whom Bridget cannot seem to stop running or help finding quietly attractive.",
+          "popularity": 4.0946,
+          "poster_path": "/dkauRl9TosBFikftrC3OVcKWDoz.jpg",
+          "release_date": "2001-04-13",
+          "title": "Bridget Jones's Diary",
+          "video": false,
+          "vote_average": 6.747,
+          "vote_count": 5261
+        }
+      ],
+      "total_pages": 10052,
+      "total_results": 201033
+    }
+  }
+  """#
