--- Tests/TMDBTests/Movies/Data+PopularMovies.swift
@@ -1,415 +1,0 @@
-import Foundation
-
-extension Data {
-  static let popularMovies = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "page": 1,
-    "results": [
-      {
-        "adult": false,
-        "backdrop_path": "/iZLqwEwUViJdSkGVjePGhxYzbDb.jpg",
-        "genre_ids": [
-          878,
-          53,
-          27,
-          18
-        ],
-        "id": 755898,
-        "original_language": "en",
-        "original_title": "War of the Worlds",
-        "overview": "Will Radford is a top cyber-security analyst for Homeland Security who tracks potential threats to national security through a mass surveillance program, until one day an attack by an unknown entity leads him to question whether the government is hiding something from him... and from the rest of the world.",
-        "popularity": 1756.9059,
-        "poster_path": "/yvirUYrva23IudARHn3mMGVxWqM.jpg",
-        "release_date": "2025-07-29",
-        "title": "War of the Worlds",
-        "video": false,
-        "vote_average": 4.532,
-        "vote_count": 95
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/8J6UlIFcU7eZfq9iCLbgc8Auklg.jpg",
-        "genre_ids": [
-          14,
-          10751,
-          28
-        ],
-        "id": 1087192,
-        "original_language": "en",
-        "original_title": "How to Train Your Dragon",
-        "overview": "On the rugged isle of Berk, where Vikings and dragons have been bitter enemies for generations, Hiccup stands apart, defying centuries of tradition when he befriends Toothless, a feared Night Fury dragon. Their unlikely bond reveals the true nature of dragons, challenging the very foundations of Viking society.",
-        "popularity": 576.648,
-        "poster_path": "/53dsJ3oEnBhTBVMigWJ9tkA5bzJ.jpg",
-        "release_date": "2025-06-06",
-        "title": "How to Train Your Dragon",
-        "video": false,
-        "vote_average": 8.057,
-        "vote_count": 1398
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/1RgPyOhN4DRs225BGTlHJqCudII.jpg",
-        "genre_ids": [
-          16,
-          28,
-          14,
-          53
-        ],
-        "id": 1311031,
-        "original_language": "ja",
-        "original_title": "劇場版「鬼滅の刃」無限城編 第一章 猗窩座再来",
-        "overview": "As the Demon Slayer Corps members and Hashira engaged in a group strength training program, the Hashira Training, in preparation for the forthcoming battle against the demons, Muzan Kibutsuji appears at the Ubuyashiki Mansion. With the head of the Demon Corps in danger, Tanjiro and the Hashira rush to the headquarters but are plunged into a deep descent to a mysterious space by the hands of Muzan Kibutsuji.  The destination of where Tanjiro and Demon Slayer Corps have fallen is the demons' stronghold – the Infinity Castle. And so, the battleground is set as the final battle between the Demon Slayer Corps and the demons ignites.",
-        "popularity": 531.3633,
-        "poster_path": "/aFRDH3P7TX61FVGpaLhKr6QiOC1.jpg",
-        "release_date": "2025-07-18",
-        "title": "Demon Slayer: Kimetsu no Yaiba Infinity Castle",
-        "video": false,
-        "vote_average": 7.12,
-        "vote_count": 54
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/shRevbFSTMZK0PheyYhZq135Kcr.jpg",
-        "genre_ids": [
-          28,
-          27,
-          10749
-        ],
-        "id": 986206,
-        "original_language": "en",
-        "original_title": "Night Carnage",
-        "overview": "A blogger who is also a werewolf meets a dashing playboy with a dark secret of his own. Starring Logan Andrews and Christian Howard.",
-        "popularity": 471.5654,
-        "poster_path": "/gep8Mk0ndUn8IXyXouPz5cvIvzj.jpg",
-        "release_date": "2025-07-29",
-        "title": "Night Carnage",
-        "video": false,
-        "vote_average": 0,
-        "vote_count": 0
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/x5dVPttNDZaVRTvbk7pYrtGZoZN.jpg",
-        "genre_ids": [
-          35
-        ],
-        "id": 1263256,
-        "original_language": "en",
-        "original_title": "Happy Gilmore 2",
-        "overview": "Happy Gilmore isn't done with golf — not by a long shot. Since his retirement after his first Tour Championship win, Gilmore returns to finance his daughter's ballet classes.",
-        "popularity": 371.1605,
-        "poster_path": "/ynT06XivgBDkg7AtbDbX1dJeBGY.jpg",
-        "release_date": "2025-07-25",
-        "title": "Happy Gilmore 2",
-        "video": false,
-        "vote_average": 6.749,
-        "vote_count": 389
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/7Zx3wDG5bBtcfk8lcnCWDOLM4Y4.jpg",
-        "genre_ids": [
-          10751,
-          878,
-          35,
-          12
-        ],
-        "id": 552524,
-        "original_language": "en",
-        "original_title": "Lilo & Stitch",
-        "overview": "The wildly funny and touching story of a lonely Hawaiian girl and the fugitive alien who helps to mend her broken family.",
-        "popularity": 372.4416,
-        "poster_path": "/tUae3mefrDVTgm5mRzqWnZK6fOP.jpg",
-        "release_date": "2025-05-17",
-        "title": "Lilo & Stitch",
-        "video": false,
-        "vote_average": 7.329,
-        "vote_count": 1217
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/cEQMqB3ahd4mfeUN6VGC0ouVnZZ.jpg",
-        "genre_ids": [
-          28,
-          878,
-          53
-        ],
-        "id": 1071585,
-        "original_language": "en",
-        "original_title": "M3GAN 2.0",
-        "overview": "After the underlying tech for M3GAN is stolen and misused by a powerful defense contractor to create a military-grade weapon known as Amelia, M3GAN's creator Gemma realizes that the only option is to resurrect M3GAN and give her a few upgrades, making her faster, stronger, and more lethal.",
-        "popularity": 364.3926,
-        "poster_path": "/oekamLQrwlJjRNmfaBE4llIvkir.jpg",
-        "release_date": "2025-06-25",
-        "title": "M3GAN 2.0",
-        "video": false,
-        "vote_average": 7.68,
-        "vote_count": 575
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/wFzI8o0jEMYfNugglP4wjFIr647.jpg",
-        "genre_ids": [
-          28,
-          53,
-          80
-        ],
-        "id": 715253,
-        "original_language": "ko",
-        "original_title": "유령",
-        "overview": "In 1933, when Korea was under Japanese occupation, five people in Gyeongseong are suspected to be \"Phantom\" spies of the anti-Japanese organization.",
-        "popularity": 331.2327,
-        "poster_path": "/n1rxZ4wovf7BYhLNwCqde8I3I2N.jpg",
-        "release_date": "2023-01-18",
-        "title": "Phantom",
-        "video": false,
-        "vote_average": 6.67,
-        "vote_count": 50
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/s94NjfKkcSczZ1FembwmQZwsuwY.jpg",
-        "genre_ids": [
-          878,
-          12
-        ],
-        "id": 617126,
-        "original_language": "en",
-        "original_title": "The Fantastic 4: First Steps",
-        "overview": "Against the vibrant backdrop of a 1960s-inspired, retro-futuristic world, Marvel's First Family is forced to balance their roles as heroes with the strength of their family bond, while defending Earth from a ravenous space god called Galactus and his enigmatic Herald, Silver Surfer.",
-        "popularity": 327.4321,
-        "poster_path": "/x26MtUlwtWD26d0G0FXcppxCJio.jpg",
-        "release_date": "2025-07-23",
-        "title": "The Fantastic 4: First Steps",
-        "video": false,
-        "vote_average": 7.354,
-        "vote_count": 721
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/h6gChZHFpmbwqwV3uQsoakp77p1.jpg",
-        "genre_ids": [
-          28,
-          35
-        ],
-        "id": 1124619,
-        "original_language": "en",
-        "original_title": "Bride Hard",
-        "overview": "Sam is a secret agent whose toughest mission to date is pleasing her bride-to-be best friend at a lavish destination wedding. When a team of mercenaries crashes the party and takes the guests hostage, Sam is thrown into a fight unlike any before — one where she can’t risk blowing her cover or ruining the big day. As she takes on the bad guys in a high-stakes battle disguised as a fairy-tale affair, she realizes the real threat might be closer than she thinks.",
-        "popularity": 266.3935,
-        "poster_path": "/3mExdWLSxAiUCb4NMcYmxSkO7n4.jpg",
-        "release_date": "2025-06-19",
-        "title": "Bride Hard",
-        "video": false,
-        "vote_average": 6.029,
-        "vote_count": 52
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/6WqqEjiycNvDLjbEClM1zCwIbDD.jpg",
-        "genre_ids": [
-          27,
-          53,
-          878
-        ],
-        "id": 1100988,
-        "original_language": "en",
-        "original_title": "28 Years Later",
-        "overview": "Twenty-eight years since the rage virus escaped a biological weapons laboratory, now, still in a ruthlessly enforced quarantine, some have found ways to exist amidst the infected. One such group lives on a small island connected to the mainland by a single, heavily-defended causeway. When one member departs on a mission into the dark heart of the mainland, he discovers secrets, wonders, and horrors that have mutated not only the infected but other survivors as well.",
-        "popularity": 403.4019,
-        "poster_path": "/mIg1qCkVxnAlM2TK3RUF0tdEXlE.jpg",
-        "release_date": "2025-06-18",
-        "title": "28 Years Later",
-        "video": false,
-        "vote_average": 7,
-        "vote_count": 808
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/eU7IfdWq8KQy0oNd4kKXS0QUR08.jpg",
-        "genre_ids": [
-          878,
-          12,
-          28
-        ],
-        "id": 1061474,
-        "original_language": "en",
-        "original_title": "Superman",
-        "overview": "Superman, a journalist in Metropolis, embarks on a journey to reconcile his Kryptonian heritage with his human upbringing as Clark Kent.",
-        "popularity": 249.4787,
-        "poster_path": "/ombsmhYUqR4qqOLOxAyr5V8hbyv.jpg",
-        "release_date": "2025-07-09",
-        "title": "Superman",
-        "video": false,
-        "vote_average": 7.424,
-        "vote_count": 1337
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/962KXsr09uK8wrmUg9TjzmE7c4e.jpg",
-        "genre_ids": [
-          28,
-          53,
-          18
-        ],
-        "id": 1119878,
-        "original_language": "en",
-        "original_title": "Ice Road: Vengeance",
-        "overview": "Big rig ice road driver Mike McCann travels to Nepal to scatter his late brother’s ashes on Mt. Everest. While on a packed tour bus traversing the deadly 12,000 ft. terrain of the infamous Road to the Sky, McCann and his mountain guide encounter a group of mercenaries and must fight to save themselves, the busload of innocent travelers, and the local villagers’ homeland.",
-        "popularity": 242.7462,
-        "poster_path": "/cQN9rZj06rXMVkk76UF1DfBAico.jpg",
-        "release_date": "2025-06-27",
-        "title": "Ice Road: Vengeance",
-        "video": false,
-        "vote_average": 6.8,
-        "vote_count": 155
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/7Q2CmqIVJuDAESPPp76rWIiA0AD.jpg",
-        "genre_ids": [
-          28,
-          12,
-          18
-        ],
-        "id": 1011477,
-        "original_language": "en",
-        "original_title": "Karate Kid: Legends",
-        "overview": "After a family tragedy, kung fu prodigy Li Fong is uprooted from his home in Beijing and forced to move to New York City with his mother. When a new friend needs his help, Li enters a karate competition – but his skills alone aren't enough. Li's kung fu teacher Mr. Han enlists original Karate Kid Daniel LaRusso for help, and Li learns a new way to fight, merging their two styles into one for the ultimate martial arts showdown.",
-        "popularity": 233.4339,
-        "poster_path": "/AEgggzRr1vZCLY86MAp93li43z.jpg",
-        "release_date": "2025-05-08",
-        "title": "Karate Kid: Legends",
-        "video": false,
-        "vote_average": 7.16,
-        "vote_count": 581
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/8or4S9BPhkeYK0vlKsPFee4JVWI.jpg",
-        "genre_ids": [
-          28,
-          18
-        ],
-        "id": 1315986,
-        "original_language": "en",
-        "original_title": "Man with No Past",
-        "overview": "Waking up in an unfamiliar city, a man with no memory must confront the mysteries of his own identity. However, his desperate search to uncover the past pits him against a powerful enemy, leading to a showdown that ultimately reveals the truth.",
-        "popularity": 205.9865,
-        "poster_path": "/eWHvROuznSzcxBAAkzX1X0Rmzoe.jpg",
-        "release_date": "2025-01-13",
-        "title": "Man with No Past",
-        "video": false,
-        "vote_average": 6.595,
-        "vote_count": 42
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/fQOV47FHTJdaSuSUNlzP3zXUZWE.jpg",
-        "genre_ids": [
-          878,
-          12,
-          28
-        ],
-        "id": 1234821,
-        "original_language": "en",
-        "original_title": "Jurassic World Rebirth",
-        "overview": "Five years after the events of Jurassic World Dominion, covert operations expert Zora Bennett is contracted to lead a skilled team on a top-secret mission to secure genetic material from the world's three most massive dinosaurs. When Zora's operation intersects with a civilian family whose boating expedition was capsized, they all find themselves stranded on an island where they come face-to-face with a sinister, shocking discovery that's been hidden from the world for decades.",
-        "popularity": 222.9024,
-        "poster_path": "/1RICxzeoNCAO5NpcRMIgg1XT6fm.jpg",
-        "release_date": "2025-07-01",
-        "title": "Jurassic World Rebirth",
-        "video": false,
-        "vote_average": 6.334,
-        "vote_count": 869
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/l3ycQYwWmbz7p8otwbomFDXIEhn.jpg",
-        "genre_ids": [
-          16,
-          14,
-          35,
-          10402,
-          10751,
-          28
-        ],
-        "id": 803796,
-        "original_language": "en",
-        "original_title": "KPop Demon Hunters",
-        "overview": "When K-pop superstars Rumi, Mira and Zoey aren't selling out stadiums, they're using their secret powers to protect their fans from supernatural threats.",
-        "popularity": 187.963,
-        "poster_path": "/22AouvwlhlXbe3nrFcjzL24bvWH.jpg",
-        "release_date": "2025-06-20",
-        "title": "KPop Demon Hunters",
-        "video": false,
-        "vote_average": 8.484,
-        "vote_count": 871
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/oPgXVSdGR9dGwbmvIToOCMmsdc2.jpg",
-        "genre_ids": [
-          28,
-          53,
-          80
-        ],
-        "id": 541671,
-        "original_language": "en",
-        "original_title": "Ballerina",
-        "overview": "Taking place during the events of John Wick: Chapter 3 – Parabellum, Eve Macarro begins her training in the assassin traditions of the Ruska Roma.",
-        "popularity": 193.5458,
-        "poster_path": "/2VUmvqsHb6cEtdfscEA6fqqVzLg.jpg",
-        "release_date": "2025-06-04",
-        "title": "Ballerina",
-        "video": false,
-        "vote_average": 7.461,
-        "vote_count": 1223
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/rthMuZfFv4fqEU4JVbgSW9wQ8rs.jpg",
-        "genre_ids": [
-          28,
-          878,
-          12
-        ],
-        "id": 986056,
-        "original_language": "en",
-        "original_title": "Thunderbolts*",
-        "overview": "After finding themselves ensnared in a death trap, seven disillusioned castoffs must embark on a dangerous mission that will force them to confront the darkest corners of their pasts.",
-        "popularity": 174.5015,
-        "poster_path": "/hqcexYHbiTBfDIdDWxrxPtVndBX.jpg",
-        "release_date": "2025-04-30",
-        "title": "Thunderbolts*",
-        "video": false,
-        "vote_average": 7.417,
-        "vote_count": 2093
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/uIpJPDNFoeX0TVml9smPrs9KUVx.jpg",
-        "genre_ids": [
-          27,
-          9648
-        ],
-        "id": 574475,
-        "original_language": "en",
-        "original_title": "Final Destination Bloodlines",
-        "overview": "Plagued by a violent recurring nightmare, college student Stefanie heads home to track down the one person who might be able to break the cycle and save her family from the grisly demise that inevitably awaits them all.",
-        "popularity": 158.9292,
-        "poster_path": "/6WxhEvFsauuACfv8HyoVX6mZKFj.jpg",
-        "release_date": "2025-05-14",
-        "title": "Final Destination Bloodlines",
-        "video": false,
-        "vote_average": 7.162,
-        "vote_count": 1803
-      }
-    ],
-    "total_pages": 51739,
-    "total_results": 1034765
-  }
-  """#
