--- Tests/TMDBTests/Movies/Data+TopRatedMovies.swift
@@ -1,411 +1,0 @@
-import Foundation
-
-extension Data {
-  static let topRatedMovies = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "page": 1,
-    "results": [
-      {
-        "adult": false,
-        "backdrop_path": "/zfbjgQE1uSd9wiPTX4VzsLi0rGG.jpg",
-        "genre_ids": [
-          18,
-          80
-        ],
-        "id": 278,
-        "original_language": "en",
-        "original_title": "The Shawshank Redemption",
-        "overview": "Imprisoned in the 1940s for the double murder of his wife and her lover, upstanding banker Andy Dufresne begins a new life at the Shawshank prison, where he puts his accounting skills to work for an amoral warden. During his long stretch in prison, Dufresne comes to be admired by the other inmates -- including an older prisoner named Red -- for his integrity and unquenchable sense of hope.",
-        "popularity": 38.0206,
-        "poster_path": "/9cqNxx0GxF0bflZmeSMuL5tnGzr.jpg",
-        "release_date": "1994-09-23",
-        "title": "The Shawshank Redemption",
-        "video": false,
-        "vote_average": 8.7,
-        "vote_count": 28616
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/htuuuEwAvDVECMpb0ltLLyZyDDt.jpg",
-        "genre_ids": [
-          18,
-          80
-        ],
-        "id": 238,
-        "original_language": "en",
-        "original_title": "The Godfather",
-        "overview": "Spanning the years 1945 to 1955, a chronicle of the fictional Italian-American Corleone crime family. When organized crime family patriarch, Vito Corleone barely survives an attempt on his life, his youngest son, Michael steps in to take care of the would-be killers, launching a campaign of bloody revenge.",
-        "popularity": 29.3323,
-        "poster_path": "/3bhkrj58Vtu7enYsRolD1fZdja1.jpg",
-        "release_date": "1972-03-14",
-        "title": "The Godfather",
-        "video": false,
-        "vote_average": 8.686,
-        "vote_count": 21659
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/kGzFbGhp99zva6oZODW5atUtnqi.jpg",
-        "genre_ids": [
-          18,
-          80
-        ],
-        "id": 240,
-        "original_language": "en",
-        "original_title": "The Godfather Part II",
-        "overview": "In the continuing saga of the Corleone crime family, a young Vito Corleone grows up in Sicily and in 1910s New York. In the 1950s, Michael Corleone attempts to expand the family business into Las Vegas, Hollywood and Cuba.",
-        "popularity": 21.6599,
-        "poster_path": "/hek3koDUyRQk7FIhPXsa6mT2Zc3.jpg",
-        "release_date": "1974-12-20",
-        "title": "The Godfather Part II",
-        "video": false,
-        "vote_average": 8.571,
-        "vote_count": 13078
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/zb6fM1CX41D9rF9hdgclu0peUmy.jpg",
-        "genre_ids": [
-          18,
-          36,
-          10752
-        ],
-        "id": 424,
-        "original_language": "en",
-        "original_title": "Schindler's List",
-        "overview": "The true story of how businessman Oskar Schindler saved over a thousand Jewish lives from the Nazis while they worked as slaves in his factory during World War II.",
-        "popularity": 20.9339,
-        "poster_path": "/sF1U4EUQS8YHUYjNl3pMGNIQyr0.jpg",
-        "release_date": "1993-12-15",
-        "title": "Schindler's List",
-        "video": false,
-        "vote_average": 8.566,
-        "vote_count": 16584
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/bxgTSUenZDHNFerQ1whRKplrMKF.jpg",
-        "genre_ids": [
-          18
-        ],
-        "id": 389,
-        "original_language": "en",
-        "original_title": "12 Angry Men",
-        "overview": "The defense and the prosecution have rested and the jury is filing into the jury room to decide if a young Spanish-American is guilty or innocent of murdering his father. What begins as an open and shut case soon becomes a mini-drama of each of the jurors' prejudices and preconceptions about the trial, the accused, and each other.",
-        "popularity": 21.9162,
-        "poster_path": "/ow3wq89wM8qd5X7hWKxiRfsFf9C.jpg",
-        "release_date": "1957-04-10",
-        "title": "12 Angry Men",
-        "video": false,
-        "vote_average": 8.549,
-        "vote_count": 9283
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/6oaL4DP75yABrd5EbC4H2zq5ghc.jpg",
-        "genre_ids": [
-          16,
-          10751,
-          14
-        ],
-        "id": 129,
-        "original_language": "ja",
-        "original_title": "千と千尋の神隠し",
-        "overview": "A young girl, Chihiro, becomes trapped in a strange new world of spirits. When her parents undergo a mysterious transformation, she must call upon the courage she never knew she had to free her family.",
-        "popularity": 23.6976,
-        "poster_path": "/39wmItIWsg5sZMyRUHLkWBcuVCM.jpg",
-        "release_date": "2001-07-20",
-        "title": "Spirited Away",
-        "video": false,
-        "vote_average": 8.536,
-        "vote_count": 17322
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/enNubozHn9pXi0ycTVYUWfpHZm.jpg",
-        "genre_ids": [
-          18,
-          28,
-          80,
-          53
-        ],
-        "id": 155,
-        "original_language": "en",
-        "original_title": "The Dark Knight",
-        "overview": "Batman raises the stakes in his war on crime. With the help of Lt. Jim Gordon and District Attorney Harvey Dent, Batman sets out to dismantle the remaining criminal organizations that plague the streets. The partnership proves to be effective, but they soon find themselves prey to a reign of chaos unleashed by a rising criminal mastermind known to the terrified citizens of Gotham as the Joker.",
-        "popularity": 32.0719,
-        "poster_path": "/qJ2tW6WMUDux911r6m7haRef0WH.jpg",
-        "release_date": "2008-07-16",
-        "title": "The Dark Knight",
-        "video": false,
-        "vote_average": 8.522,
-        "vote_count": 34150
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/90ez6ArvpO8bvpyIngBuwXOqJm5.jpg",
-        "genre_ids": [
-          35,
-          18,
-          10749
-        ],
-        "id": 19404,
-        "original_language": "hi",
-        "original_title": "दिलवाले दुल्हनिया ले जायेंगे",
-        "overview": "Raj is a rich, carefree, happy-go-lucky second generation NRI. Simran is the daughter of Chaudhary Baldev Singh, who in spite of being an NRI is very strict about adherence to Indian values. Simran has left for India to be married to her childhood fiancé. Raj leaves for India with a mission at his hands, to claim his lady love under the noses of her whole family. Thus begins a saga.",
-        "popularity": 8.899,
-        "poster_path": "/2CAL2433ZeIihfX1Hb2139CX0pW.jpg",
-        "release_date": "1995-10-20",
-        "title": "Dilwale Dulhania Le Jayenge",
-        "video": false,
-        "vote_average": 8.5,
-        "vote_count": 4498
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/vxJ08SvwomfKbpboCWynC3uqUg4.jpg",
-        "genre_ids": [
-          14,
-          18,
-          80
-        ],
-        "id": 497,
-        "original_language": "en",
-        "original_title": "The Green Mile",
-        "overview": "A supernatural tale set on death row in a Southern prison, where gentle giant John Coffey possesses the mysterious power to heal people's ailments. When the cell block's head guard, Paul Edgecomb, recognizes Coffey's miraculous gift, he tries desperately to help stave off the condemned man's execution.",
-        "popularity": 21.276,
-        "poster_path": "/8VG8fDNiy50H4FedGwdSVUPoaJe.jpg",
-        "release_date": "1999-12-10",
-        "title": "The Green Mile",
-        "video": false,
-        "vote_average": 8.504,
-        "vote_count": 18238
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/hiKmpZMGZsrkA3cdce8a7Dpos1j.jpg",
-        "genre_ids": [
-          35,
-          53,
-          18
-        ],
-        "id": 496243,
-        "original_language": "ko",
-        "original_title": "기생충",
-        "overview": "All unemployed, Ki-taek's family takes peculiar interest in the wealthy and glamorous Parks for their livelihood until they get entangled in an unexpected incident.",
-        "popularity": 28.6702,
-        "poster_path": "/7IiTTgloJzvGI1TAYymCfbfl3vT.jpg",
-        "release_date": "2019-05-30",
-        "title": "Parasite",
-        "video": false,
-        "vote_average": 8.499,
-        "vote_count": 19306
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/2u7zbn8EudG6kLlBzUYqP8RyFU4.jpg",
-        "genre_ids": [
-          12,
-          14,
-          28
-        ],
-        "id": 122,
-        "original_language": "en",
-        "original_title": "The Lord of the Rings: The Return of the King",
-        "overview": "As armies mass for a final battle that will decide the fate of the world--and powerful, ancient forces of Light and Dark compete to determine the outcome--one member of the Fellowship of the Ring is revealed as the noble heir to the throne of the Kings of Men. Yet, the sole hope for triumph over evil lies with a brave hobbit, Frodo, who, accompanied by his loyal friend Sam and the hideous, wretched Gollum, ventures deep into the very dark heart of Mordor on his seemingly impossible quest to destroy the Ring of Power.​",
-        "popularity": 26.6628,
-        "poster_path": "/rCzpDGLbOoPwLjy3OAm5NUPOTrC.jpg",
-        "release_date": "2003-12-17",
-        "title": "The Lord of the Rings: The Return of the King",
-        "video": false,
-        "vote_average": 8.488,
-        "vote_count": 25296
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/suaEOtk1N1sgg2MTM7oZd2cfVp3.jpg",
-        "genre_ids": [
-          53,
-          80,
-          35,
-          18
-        ],
-        "id": 680,
-        "original_language": "en",
-        "original_title": "Pulp Fiction",
-        "overview": "A burger-loving hit man, his philosophical partner, a drug-addled gangster's moll and a washed-up boxer converge in this sprawling, comedic crime caper. Their adventures unfurl in three stories that ingeniously trip back and forth in time.",
-        "popularity": 26.5225,
-        "poster_path": "/vQWk5YBFWF4bZaofAbv0tShwBvQ.jpg",
-        "release_date": "1994-09-10",
-        "title": "Pulp Fiction",
-        "video": false,
-        "vote_average": 8.5,
-        "vote_count": 28911
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/8x9iKH8kWA0zdkgNdpAew7OstYe.jpg",
-        "genre_ids": [
-          16,
-          10749,
-          18
-        ],
-        "id": 372058,
-        "original_language": "ja",
-        "original_title": "君の名は。",
-        "overview": "High schoolers Mitsuha and Taki are complete strangers living separate lives. But one night, they suddenly switch places. Mitsuha wakes up in Taki’s body, and he in hers. This bizarre occurrence continues to happen randomly, and the two must adjust their lives around each other.",
-        "popularity": 16.2736,
-        "poster_path": "/q719jXXEzOoYaps6babgKnONONX.jpg",
-        "release_date": "2016-08-26",
-        "title": "Your Name.",
-        "video": false,
-        "vote_average": 8.482,
-        "vote_count": 11890
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/l3ycQYwWmbz7p8otwbomFDXIEhn.jpg",
-        "genre_ids": [
-          16,
-          14,
-          35,
-          10402,
-          10751,
-          28
-        ],
-        "id": 803796,
-        "original_language": "en",
-        "original_title": "KPop Demon Hunters",
-        "overview": "When K-pop superstars Rumi, Mira and Zoey aren't selling out stadiums, they're using their secret powers to protect their fans from supernatural threats.",
-        "popularity": 187.963,
-        "poster_path": "/22AouvwlhlXbe3nrFcjzL24bvWH.jpg",
-        "release_date": "2025-06-20",
-        "title": "KPop Demon Hunters",
-        "video": false,
-        "vote_average": 8.478,
-        "vote_count": 875
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/67HggiWaP9ZLv5sPYmyRV37yAJM.jpg",
-        "genre_ids": [
-          35,
-          18,
-          10749
-        ],
-        "id": 13,
-        "original_language": "en",
-        "original_title": "Forrest Gump",
-        "overview": "A man with a low IQ has accomplished great things in his life and been present during significant historic events—in each case, far exceeding what anyone imagined he could do. But despite all he has achieved, his one true love eludes him.",
-        "popularity": 23.1311,
-        "poster_path": "/arw2vcBveWOVZr6pxd9XTd1TdQa.jpg",
-        "release_date": "1994-06-23",
-        "title": "Forrest Gump",
-        "video": false,
-        "vote_average": 8.468,
-        "vote_count": 28441
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/x4biAVdPVCghBlsVIzB6NmbghIz.jpg",
-        "genre_ids": [
-          37
-        ],
-        "id": 429,
-        "original_language": "it",
-        "original_title": "Il buono, il brutto, il cattivo",
-        "overview": "While the Civil War rages on between the Union and the Confederacy, three men – a quiet loner, a ruthless hitman, and a Mexican bandit – comb the American Southwest in search of a strongbox containing $200,000 in stolen gold.",
-        "popularity": 19.3186,
-        "poster_path": "/bX2xnavhMYjWDoZp1VM6VnU1xwe.jpg",
-        "release_date": "1966-12-22",
-        "title": "The Good, the Bad and the Ugly",
-        "video": false,
-        "vote_average": 8.466,
-        "vote_count": 9097
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/vgnoBSVzWAV9sNQUORaDGvDp7wx.jpg",
-        "genre_ids": [
-          12,
-          18,
-          878
-        ],
-        "id": 157336,
-        "original_language": "en",
-        "original_title": "Interstellar",
-        "overview": "The adventures of a group of explorers who make use of a newly discovered wormhole to surpass the limitations on human space travel and conquer the vast distances involved in an interstellar voyage.",
-        "popularity": 49.4336,
-        "poster_path": "/gEU2QniE6E77NI6lCU6MxlNBvIx.jpg",
-        "release_date": "2014-11-05",
-        "title": "Interstellar",
-        "video": false,
-        "vote_average": 8.458,
-        "vote_count": 37543
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/qvZ91FwMq6O47VViAr8vZNQz3WI.jpg",
-        "genre_ids": [
-          28,
-          18
-        ],
-        "id": 346,
-        "original_language": "ja",
-        "original_title": "七人の侍",
-        "overview": "A samurai answers a village's request for protection after he falls on hard times. The town needs protection from bandits, so the samurai gathers six others to help him teach the people how to defend themselves, and the villagers provide the soldiers with food.",
-        "popularity": 11.2911,
-        "poster_path": "/lOMGc8bnSwQhS4XyE1S99uH8NXf.jpg",
-        "release_date": "1954-04-26",
-        "title": "Seven Samurai",
-        "video": false,
-        "vote_average": 8.455,
-        "vote_count": 3950
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/7TF4p86ZafnxFuNqWdhpHXFO244.jpg",
-        "genre_ids": [
-          18,
-          80
-        ],
-        "id": 769,
-        "original_language": "en",
-        "original_title": "GoodFellas",
-        "overview": "The true story of Henry Hill, a half-Irish, half-Sicilian Brooklyn kid who is adopted by neighbourhood gangsters at an early age and climbs the ranks of a Mafia family under the guidance of Jimmy Conway.",
-        "popularity": 14.4579,
-        "poster_path": "/aKuFiU82s5ISJpGZp7YkIr3kCUd.jpg",
-        "release_date": "1990-09-12",
-        "title": "GoodFellas",
-        "video": false,
-        "vote_average": 8.454,
-        "vote_count": 13506
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/dlC0ed9Ugh3FzydnkBtV5lRXUu4.jpg",
-        "genre_ids": [
-          16,
-          18,
-          10752
-        ],
-        "id": 12477,
-        "original_language": "ja",
-        "original_title": "火垂るの墓",
-        "overview": "In the final months of World War II, 14-year-old Seita and his sister Setsuko are orphaned when their mother is killed during an air raid in Kobe, Japan. After a falling out with their aunt, they move into an abandoned bomb shelter. With no surviving relatives and their emergency rations depleted, Seita and Setsuko struggle to survive.",
-        "popularity": 0.0097,
-        "poster_path": "/k9tv1rXZbOhH7eiCk378x61kNQ1.jpg",
-        "release_date": "1988-04-16",
-        "title": "Grave of the Fireflies",
-        "video": false,
-        "vote_average": 8.45,
-        "vote_count": 5982
-      }
-    ],
-    "total_pages": 514,
-    "total_results": 10267
-  }
-  """#
