--- Tests/TMDBTests/Movies/Data+TrendingMovies.swift
@@ -1,433 +1,0 @@
-import Foundation
-
-extension Data {
-  static let trendingMovies = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "page": 1,
-    "results": [
-      {
-        "adult": false,
-        "backdrop_path": "/zxi6WQPVc0uQAG5TtLsKvxYHApC.jpg",
-        "id": 980477,
-        "title": "Ne Zha 2",
-        "original_title": "哪吒之魔童闹海",
-        "overview": "A rebellious young boy, Ne Zha, is feared by the gods and born to mortal parents with wild, uncontrolled powers. Now he's faced with an ancient force intent on destroying humanity, he must grow up to become the hero the world needs.",
-        "poster_path": "/5lUmWTGkEcYnXujixXn31o9q2T0.jpg",
-        "media_type": "movie",
-        "original_language": "zh",
-        "genre_ids": [
-          16,
-          14,
-          12,
-          28
-        ],
-        "popularity": 104.1626,
-        "release_date": "2025-01-29",
-        "video": false,
-        "vote_average": 8.098,
-        "vote_count": 200
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/A466i5iATrpbVjX30clP1Zyfp31.jpg",
-        "id": 1307078,
-        "title": "My Oxford Year",
-        "original_title": "My Oxford Year",
-        "overview": "An ambitious American fulfilling her dream of studying at Oxford falls for a charming Brit hiding a secret that may upend her perfectly planned life.",
-        "poster_path": "/jrhXbIOFingzdLjkccjg9vZnqIp.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          10749,
-          35,
-          18
-        ],
-        "popularity": 7.3366,
-        "release_date": "2025-07-31",
-        "video": false,
-        "vote_average": 7.712,
-        "vote_count": 26
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/6WqqEjiycNvDLjbEClM1zCwIbDD.jpg",
-        "id": 1100988,
-        "title": "28 Years Later",
-        "original_title": "28 Years Later",
-        "overview": "Twenty-eight years since the rage virus escaped a biological weapons laboratory, now, still in a ruthlessly enforced quarantine, some have found ways to exist amidst the infected. One such group lives on a small island connected to the mainland by a single, heavily-defended causeway. When one member departs on a mission into the dark heart of the mainland, he discovers secrets, wonders, and horrors that have mutated not only the infected but other survivors as well.",
-        "poster_path": "/mIg1qCkVxnAlM2TK3RUF0tdEXlE.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          27,
-          53,
-          878
-        ],
-        "popularity": 403.4019,
-        "release_date": "2025-06-18",
-        "video": false,
-        "vote_average": 7,
-        "vote_count": 811
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/iZLqwEwUViJdSkGVjePGhxYzbDb.jpg",
-        "id": 755898,
-        "title": "War of the Worlds",
-        "original_title": "War of the Worlds",
-        "overview": "Will Radford is a top cyber-security analyst for Homeland Security who tracks potential threats to national security through a mass surveillance program, until one day an attack by an unknown entity leads him to question whether the government is hiding something from him... and from the rest of the world.",
-        "poster_path": "/yvirUYrva23IudARHn3mMGVxWqM.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          878,
-          53,
-          27,
-          18
-        ],
-        "popularity": 1756.9059,
-        "release_date": "2025-07-29",
-        "video": false,
-        "vote_average": 4.556,
-        "vote_count": 99
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/s94NjfKkcSczZ1FembwmQZwsuwY.jpg",
-        "id": 617126,
-        "title": "The Fantastic 4: First Steps",
-        "original_title": "The Fantastic 4: First Steps",
-        "overview": "Against the vibrant backdrop of a 1960s-inspired, retro-futuristic world, Marvel's First Family is forced to balance their roles as heroes with the strength of their family bond, while defending Earth from a ravenous space god called Galactus and his enigmatic Herald, Silver Surfer.",
-        "poster_path": "/x26MtUlwtWD26d0G0FXcppxCJio.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          878,
-          12
-        ],
-        "popularity": 327.4321,
-        "release_date": "2025-07-23",
-        "video": false,
-        "vote_average": 7.4,
-        "vote_count": 727
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/fUoHnP40Mpuo14jgAOMJMPXSVfG.jpg",
-        "id": 1035259,
-        "title": "The Naked Gun",
-        "original_title": "The Naked Gun",
-        "overview": "Only one man has the particular set of skills... to lead Police Squad and save the world: Lt. Frank Drebin Jr.",
-        "poster_path": "/aq0JMbmSfPwG8JvAzExJPrBHqmG.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          35,
-          80
-        ],
-        "popularity": 76.5874,
-        "release_date": "2025-07-30",
-        "video": false,
-        "vote_average": 6.822,
-        "vote_count": 45
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/rthMuZfFv4fqEU4JVbgSW9wQ8rs.jpg",
-        "id": 986056,
-        "title": "Thunderbolts*",
-        "original_title": "Thunderbolts*",
-        "overview": "After finding themselves ensnared in a death trap, seven disillusioned castoffs must embark on a dangerous mission that will force them to confront the darkest corners of their pasts.",
-        "poster_path": "/hqcexYHbiTBfDIdDWxrxPtVndBX.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          28,
-          878,
-          12
-        ],
-        "popularity": 174.5015,
-        "release_date": "2025-04-30",
-        "video": false,
-        "vote_average": 7.413,
-        "vote_count": 2097
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/eU7IfdWq8KQy0oNd4kKXS0QUR08.jpg",
-        "id": 1061474,
-        "title": "Superman",
-        "original_title": "Superman",
-        "overview": "Superman, a journalist in Metropolis, embarks on a journey to reconcile his Kryptonian heritage with his human upbringing as Clark Kent.",
-        "poster_path": "/ombsmhYUqR4qqOLOxAyr5V8hbyv.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          878,
-          12,
-          28
-        ],
-        "popularity": 249.4787,
-        "release_date": "2025-07-09",
-        "video": false,
-        "vote_average": 7.4,
-        "vote_count": 1341
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/x5dVPttNDZaVRTvbk7pYrtGZoZN.jpg",
-        "id": 1263256,
-        "title": "Happy Gilmore 2",
-        "original_title": "Happy Gilmore 2",
-        "overview": "Happy Gilmore isn't done with golf — not by a long shot. Since his retirement after his first Tour Championship win, Gilmore returns to finance his daughter's ballet classes.",
-        "poster_path": "/ynT06XivgBDkg7AtbDbX1dJeBGY.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          35
-        ],
-        "popularity": 371.1605,
-        "release_date": "2025-07-25",
-        "video": false,
-        "vote_average": 6.749,
-        "vote_count": 393
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/A5V1RIXR277KuLFEa5fdu5wbiEp.jpg",
-        "id": 1187436,
-        "title": "An Honest Life",
-        "original_title": "Ett ärligt liv",
-        "overview": "When a cynical law student in search of inspiration joins a band of anarchists, he becomes entangled in a web of crime where he is more pawn than player.",
-        "poster_path": "/4hveI7tEISN8KU60UFJw1JRYKvv.jpg",
-        "media_type": "movie",
-        "original_language": "sv",
-        "genre_ids": [
-          53,
-          18
-        ],
-        "popularity": 49.0611,
-        "release_date": "2025-07-30",
-        "video": false,
-        "vote_average": 4.5,
-        "vote_count": 5
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/8J6UlIFcU7eZfq9iCLbgc8Auklg.jpg",
-        "id": 1087192,
-        "title": "How to Train Your Dragon",
-        "original_title": "How to Train Your Dragon",
-        "overview": "On the rugged isle of Berk, where Vikings and dragons have been bitter enemies for generations, Hiccup stands apart, defying centuries of tradition when he befriends Toothless, a feared Night Fury dragon. Their unlikely bond reveals the true nature of dragons, challenging the very foundations of Viking society.",
-        "poster_path": "/53dsJ3oEnBhTBVMigWJ9tkA5bzJ.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          14,
-          10751,
-          28
-        ],
-        "popularity": 576.648,
-        "release_date": "2025-06-06",
-        "video": false,
-        "vote_average": 8.045,
-        "vote_count": 1403
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/7Zx3wDG5bBtcfk8lcnCWDOLM4Y4.jpg",
-        "id": 552524,
-        "title": "Lilo & Stitch",
-        "original_title": "Lilo & Stitch",
-        "overview": "The wildly funny and touching story of a lonely Hawaiian girl and the fugitive alien who helps to mend her broken family.",
-        "poster_path": "/tUae3mefrDVTgm5mRzqWnZK6fOP.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          10751,
-          878,
-          35,
-          12
-        ],
-        "popularity": 372.4416,
-        "release_date": "2025-05-17",
-        "video": false,
-        "vote_average": 7.3,
-        "vote_count": 1219
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/jvpkBenB6hv19WWYVlaiow8zklq.jpg",
-        "id": 1175942,
-        "title": "The Bad Guys 2",
-        "original_title": "The Bad Guys 2",
-        "overview": "The now-reformed Bad Guys are trying (very, very hard) to be good, but instead find themselves hijacked into a high-stakes, globe-trotting heist, masterminded by a new team of criminals they never saw coming: The Bad Girls.",
-        "poster_path": "/26oSPnq0ct59l07QOXZKyzsiRtN.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          10751,
-          35,
-          80,
-          12,
-          16
-        ],
-        "popularity": 99.2808,
-        "release_date": "2025-07-24",
-        "video": false,
-        "vote_average": 7.3,
-        "vote_count": 22
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/lSbblLngbeZIn6G4WXDcyQ6SLhw.jpg",
-        "id": 911430,
-        "title": "F1",
-        "original_title": "F1",
-        "overview": "Racing legend Sonny Hayes is coaxed out of retirement to lead a struggling Formula 1 team—and mentor a young hotshot driver—while chasing one more chance at glory.",
-        "poster_path": "/6H6p82aWQFEKEuVUiZll6JxV8Ft.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          28,
-          18
-        ],
-        "popularity": 120.8856,
-        "release_date": "2025-06-25",
-        "video": false,
-        "vote_average": 7.6,
-        "vote_count": 949
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/lecBUG6Hsw7pYjTNAPBUgouDfjW.jpg",
-        "id": 842924,
-        "title": "The Life of Chuck",
-        "original_title": "The Life of Chuck",
-        "overview": "In this extraordinary story of an ordinary man, Charles 'Chuck' Krantz experiences the wonder of love, the heartbreak of loss, and the multitudes contained in all of us.",
-        "poster_path": "/oumprkO9bThExP8NwxBIBnvBu2v.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          18,
-          14
-        ],
-        "popularity": 62.6564,
-        "release_date": "2025-06-05",
-        "video": false,
-        "vote_average": 7.355,
-        "vote_count": 166
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/a6bItEVaxgphpMswho1wVRerv4r.jpg",
-        "id": 7451,
-        "title": "xXx",
-        "original_title": "xXx",
-        "overview": "Xander Cage is your standard adrenaline junkie with no fear and a lousy attitude. When the US Government \"recruits\" him to go on a mission, he's not exactly thrilled. His mission: to gather information on an organization that may just be planning the destruction of the world, led by the nihilistic Yorgi.",
-        "poster_path": "/xeEw3eLeSFmJgXZzmF2Efww0q3s.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          28,
-          12,
-          53,
-          80
-        ],
-        "popularity": 126.9205,
-        "release_date": "2002-08-09",
-        "video": false,
-        "vote_average": 6,
-        "vote_count": 4493
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/cEQMqB3ahd4mfeUN6VGC0ouVnZZ.jpg",
-        "id": 1071585,
-        "title": "M3GAN 2.0",
-        "original_title": "M3GAN 2.0",
-        "overview": "After the underlying tech for M3GAN is stolen and misused by a powerful defense contractor to create a military-grade weapon known as Amelia, M3GAN's creator Gemma realizes that the only option is to resurrect M3GAN and give her a few upgrades, making her faster, stronger, and more lethal.",
-        "poster_path": "/oekamLQrwlJjRNmfaBE4llIvkir.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          28,
-          878,
-          53
-        ],
-        "popularity": 364.3926,
-        "release_date": "2025-06-25",
-        "video": false,
-        "vote_average": 7.692,
-        "vote_count": 574
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/h6gChZHFpmbwqwV3uQsoakp77p1.jpg",
-        "id": 1124619,
-        "title": "Bride Hard",
-        "original_title": "Bride Hard",
-        "overview": "Sam is a secret agent whose toughest mission to date is pleasing her bride-to-be best friend at a lavish destination wedding. When a team of mercenaries crashes the party and takes the guests hostage, Sam is thrown into a fight unlike any before — one where she can’t risk blowing her cover or ruining the big day. As she takes on the bad guys in a high-stakes battle disguised as a fairy-tale affair, she realizes the real threat might be closer than she thinks.",
-        "poster_path": "/3mExdWLSxAiUCb4NMcYmxSkO7n4.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          28,
-          35
-        ],
-        "popularity": 266.3935,
-        "release_date": "2025-06-19",
-        "video": false,
-        "vote_average": 6,
-        "vote_count": 52
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/63JYqhefGrbVg0YqZ3xWqn0pwT2.jpg",
-        "id": 969681,
-        "title": "Spider-Man: Brand New Day",
-        "original_title": "Spider-Man: Brand New Day",
-        "overview": "The fourth installment in the Spider-Man franchise and part of Phase Six set in the Marvel Cinematic Universe (MCU). Plot TBA.",
-        "poster_path": "/9JCQtDCSpPR2ld55yNlEg1VwcQo.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          878,
-          28,
-          12
-        ],
-        "popularity": 7.9331,
-        "release_date": "2026-07-29",
-        "video": false,
-        "vote_average": 0,
-        "vote_count": 0
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/oPgXVSdGR9dGwbmvIToOCMmsdc2.jpg",
-        "id": 541671,
-        "title": "Ballerina",
-        "original_title": "Ballerina",
-        "overview": "Taking place during the events of John Wick: Chapter 3 – Parabellum, Eve Macarro begins her training in the assassin traditions of the Ruska Roma.",
-        "poster_path": "/2VUmvqsHb6cEtdfscEA6fqqVzLg.jpg",
-        "media_type": "movie",
-        "original_language": "en",
-        "genre_ids": [
-          28,
-          53,
-          80
-        ],
-        "popularity": 193.5458,
-        "release_date": "2025-06-04",
-        "video": false,
-        "vote_average": 7.46,
-        "vote_count": 1225
-      }
-    ],
-    "total_pages": 500,
-    "total_results": 10000
-  }
-  """#
