--- Tests/TMDBTests/TVShows/Data+AiringTodayTVShows.swift
@@ -1,448 +1,0 @@
-import Foundation
-
-extension Data {
-  static let airingTodayTVShows = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "page": 1,
-    "results": [
-      {
-        "adult": false,
-        "backdrop_path": "/sKlF9YrVu84DYMDAUZEZDCvDxK2.jpg",
-        "genre_ids": [
-          16,
-          35,
-          18,
-          10765
-        ],
-        "id": 244808,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "ぬきたし THE ANIMATION",
-        "overview": "On Seiran Island, secret society NLNS is united by Junnosuke Tachibana, who is determined to crush a law. They have found a benefactor, an unknown rich man, who offers them a secret base and cash. Their mission is to locate a mysterious girl and stop the law, keeping the SHO off their trail.",
-        "popularity": 403.4836,
-        "poster_path": "/zlybZJXfu4IpPnWXhnRtobl9BgZ.jpg",
-        "first_air_date": "2025-07-19",
-        "name": "Nukitashi the Animation",
-        "vote_average": 6.542,
-        "vote_count": 12
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/a8a8z4aJr40zYgRMfmSESppX2eO.jpg",
-        "genre_ids": [
-          10764,
-          35
-        ],
-        "id": 70672,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "아는 형님",
-        "overview": "Male celebs play make-believe as high schoolers, welcoming star transfer students every week and engaging in battles of witty humor and slapstick.",
-        "popularity": 321.8921,
-        "poster_path": "/2jIi55JtYKJTL1km8qHMuUilOWo.jpg",
-        "first_air_date": "2015-12-05",
-        "name": "Men on a Mission",
-        "vote_average": 7.528,
-        "vote_count": 36
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/5546GF3zT3yA25ydCz0KlujjRA3.jpg",
-        "genre_ids": [
-          16,
-          80,
-          9648,
-          35
-        ],
-        "id": 30983,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "名探偵コナン",
-        "overview": "The son of a world famous mystery writer, Jimmy Kudo, has achieved his own notoriety by assisting the local police as a student detective. He has always been able to solve the most difficult of criminal cases using his wits and power of reason.",
-        "popularity": 198.2992,
-        "poster_path": "/1o2o7RELi8rtRrTh8iHYkprpDrn.jpg",
-        "first_air_date": "1996-01-08",
-        "name": "Case Closed",
-        "vote_average": 8,
-        "vote_count": 704
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/9DrORnlwnjt5SeJKcuE5sqOoIdV.jpg",
-        "genre_ids": [
-          16,
-          10759,
-          10765
-        ],
-        "id": 207468,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "怪獣８号",
-        "overview": "In a world plagued by creatures known as Kaiju, Kafka Hibino aspired to enlist in The Defense Force. He makes a promise to enlist with his childhood friend, Mina Ashiro. Soon, life takes them in separate ways. While employed cleaning up after Kaiju battles, Kafka meets Reno Ichikawa. Reno's determination to join The Defense Force reawakens Kafka's promise to join Mina and protect humanity.",
-        "popularity": 94.7975,
-        "poster_path": "/PDe8WZHxH6XxM6Buxoo99aQrgw.jpg",
-        "first_air_date": "2024-04-13",
-        "name": "Kaiju No. 8",
-        "vote_average": 8.559,
-        "vote_count": 577
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/pvOI0jUuBVigqNTXClGGhIK2dSk.jpg",
-        "genre_ids": [
-          10764
-        ],
-        "id": 154770,
-        "origin_country": [
-          "CN"
-        ],
-        "original_language": "zh",
-        "original_name": "你好，星期六",
-        "overview": "\"Hello, Saturday\" is the rebranding of popular Chinese show \"Happy Camp\", hosted by He Jiong and with a fixed group of regular celebrity members.\n\nThe program showcases the talent of various stars as well as social topics through creative and diverse content and games. It aims to spread positive energy whilst keeping up with the current trends and leading youth culture.",
-        "popularity": 90.6297,
-        "poster_path": "/wsTjWy6NE2s3h0neChr0D92ddwI.jpg",
-        "first_air_date": "2022-01-01",
-        "name": "Hello Saturday",
-        "vote_average": 6.7,
-        "vote_count": 6
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/tOpyQG9MAnFN4YiDgyk5NnCGa7X.jpg",
-        "genre_ids": [
-          18,
-          10765
-        ],
-        "id": 243224,
-        "origin_country": [
-          "CN"
-        ],
-        "original_language": "zh",
-        "original_name": "凡人修仙传",
-        "overview": "Han Li, an ordinary, poor boy from a mountain village, joins a small martial arts sect by chance. Despite his mediocre talent, he eventually cultivates himself into an immortal through hard work and careful planning.",
-        "popularity": 89.4534,
-        "poster_path": "/kG8YooBxqX5BByifbA0wYeVVRYe.jpg",
-        "first_air_date": "2025-07-27",
-        "name": "The Immortal Ascension",
-        "vote_average": 6.667,
-        "vote_count": 12
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/aFBgDwXLxdbkb1ULUG2FZiuqtj1.jpg",
-        "genre_ids": [
-          18,
-          80
-        ],
-        "id": 252568,
-        "origin_country": [
-          "CN"
-        ],
-        "original_language": "zh",
-        "original_name": "暗潮缉凶",
-        "overview": "At the former Xuri Factory, a dismembered female body is discovered in an abandoned sewage pit. Veteran detective Lian Haiping and rookie officer Shi Qiangfeng, despite their differences, must work together to solve the case. Their investigation reveals unexpected social issues and brings to light an old case that complicates their personal and professional lives.",
-        "popularity": 85.0999,
-        "poster_path": "/yGZwGbgoZGYx6aBAhQIKXZ4Xq1y.jpg",
-        "first_air_date": "2025-07-31",
-        "name": "Beneath the Undertow",
-        "vote_average": 0,
-        "vote_count": 0
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/3Wi5mEdCsvApYQZh34gMOHiOjvH.jpg",
-        "genre_ids": [
-          16,
-          10759,
-          10765
-        ],
-        "id": 79481,
-        "origin_country": [
-          "CN"
-        ],
-        "original_language": "zh",
-        "original_name": "斗破苍穹",
-        "overview": "In a land where no magic is present. A land where the strong make the rules and weak have to obey. A land filled with alluring treasures and beauty, yet also filled with unforeseen danger. Three years ago, Xiao Yan, who had shown talents none had seen in decades, suddenly lost everything. His powers, his reputation, and his promise to his mother. What sorcery has caused him to lose all of his powers? And why has his fiancee suddenly shown up?",
-        "popularity": 82.379,
-        "poster_path": "/a9bJPlezCXF6u0siDxf7cZaSXaJ.jpg",
-        "first_air_date": "2017-01-07",
-        "name": "Fights Break Sphere",
-        "vote_average": 8.3,
-        "vote_count": 47
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/hF1LFAol3GaubZ6YkDR6AQeJz7q.jpg",
-        "genre_ids": [
-          16,
-          10759,
-          10765
-        ],
-        "id": 152271,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "ニンジャラ",
-        "overview": "After years of experiments, three researchers finally create Ninja-Gum. This substance awakens dormant power within people descended from ninjas, but its creation isn’t without consequences and will lead them to the mysterious Ninjala Tournament.",
-        "popularity": 66.2221,
-        "poster_path": "/2EkTiYPBvW3R1221Sk4436PkoHt.jpg",
-        "first_air_date": "2022-01-08",
-        "name": "Ninjala the Animation",
-        "vote_average": 7.667,
-        "vote_count": 6
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/s4KnYaLU3gor1wzTGQYSfXtXYGU.jpg",
-        "genre_ids": [
-          16,
-          10759
-        ],
-        "id": 222608,
-        "origin_country": [
-          "CN"
-        ],
-        "original_language": "zh",
-        "original_name": "仙武传",
-        "overview": "As a loyal disciple, Ye Chen dedicated himself to guard the spiritual medicine field for his sect. But, during a fight with enemies, the spiritual field was destroyed. His loyalty and dedication to the sect could not save him. The loyalty he thought he had obtained from his peers and lover, could not save him from betrayal. Thus, he was shamelessly banished from the sect. With the help of a flame falling from heaven, Ye Chen began to develop himself into a stronger cultivator. Battled against his opponents, unfolded his legendary life and rewrote his own story…",
-        "popularity": 71.4766,
-        "poster_path": "/mt1pH8WGzYVhfhiNmAjHcNB1HCw.jpg",
-        "first_air_date": "2023-03-19",
-        "name": "Legend of Xianwu",
-        "vote_average": 8.4,
-        "vote_count": 7
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/zcQRmnySOmVqT6WRmdztbiA9vVk.jpg",
-        "genre_ids": [
-          18
-        ],
-        "id": 67192,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "New Japan Pro Wrestling",
-        "overview": "New Japan Pro Wrestling is a Japanese promotion founded by Antonio Inoki in 1972. Owing to its TV program aired on TV Asahi, it is the largest wrestling promotion in Japan and one of the largest in the world.",
-        "popularity": 67.5547,
-        "poster_path": "/c6IZc3avqSgTzSXrrj9vV0EwRp.jpg",
-        "first_air_date": "1972-03-06",
-        "name": "New Japan Pro Wrestling",
-        "vote_average": 8,
-        "vote_count": 5
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/ullfy98STkno11rDGi44SK6llwx.jpg",
-        "genre_ids": [
-          16,
-          35
-        ],
-        "id": 96316,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "彼女、お借りします",
-        "overview": "Kinoshita Kazuya is a 20-year-old failure of a college student. He managed to kiss his girlfriend once, but was dumped after a month. Completely spiteful, Kazuya uses a certain method to date a girl. He goes to their meeting place and a beautiful girl brushing her long, black hair behind her ear was there, smiling at him. Her name was Mizuhara Chizuru. Something real is born after just a single rental! A reckless rom-com filled with love and excitement is about to begin!",
-        "popularity": 62.5104,
-        "poster_path": "/6ZpDPUNtVw6UdJoStvVlRZ62yAi.jpg",
-        "first_air_date": "2020-07-11",
-        "name": "Rent-a-Girlfriend",
-        "vote_average": 8.4,
-        "vote_count": 906
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/AdWWrcYKNXlpeWNJlmK4HCJ6Dj9.jpg",
-        "genre_ids": [
-          16,
-          9648,
-          10765
-        ],
-        "id": 232230,
-        "origin_country": [
-          "CN"
-        ],
-        "original_language": "zh",
-        "original_name": "诡秘之主",
-        "overview": "In a Victorian world of steam, dreadnoughts, and occult horrors, Zhou Mingrui awakens as Klein Moretti. He walks a razor’s edge between light and darkness, entangled with warring Churches. This is the legend of unlimited potential…and unspeakable danger.",
-        "popularity": 63.2626,
-        "poster_path": "/c8fHePq3yTn3WvZd4hupkHwsjm5.jpg",
-        "first_air_date": "2025-06-28",
-        "name": "Lord of Mysteries",
-        "vote_average": 9.256,
-        "vote_count": 39
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/nhgEr8MwQBUsttC4fn5SeUEt6v1.jpg",
-        "genre_ids": [
-          10765,
-          10759,
-          35,
-          18
-        ],
-        "id": 254002,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Critical Role",
-        "overview": "Critical Role is a weekly livestreamed show that uses roleplaying game mechanics as a means to explore and develop stories from the vast fantasy world of Exandria, with sweeping narratives intricately woven through collaboration between Game Master Matthew Mercer and his fellow cast of veteran voice actors, including Ashley Johnson, Marisha Ray, Taliesin Jaffe, Travis Willingham, Sam Riegel, Laura Bailey, and Liam O’Brien.",
-        "popularity": 65.4555,
-        "poster_path": "/keSNwdzmSgQzluKmqTI6kwDOFa7.jpg",
-        "first_air_date": "2015-03-12",
-        "name": "Critical Role",
-        "vote_average": 8.2,
-        "vote_count": 17
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/iGJ9eOBsWcOPOQkZmH8LwWub4eL.jpg",
-        "genre_ids": [
-          16,
-          10759,
-          10765
-        ],
-        "id": 220850,
-        "origin_country": [
-          "CN"
-        ],
-        "original_language": "zh",
-        "original_name": "炼气十万年",
-        "overview": "One hundred thousand years ago, the Heavenly Lan Clan reigned over the cultivation world, and all the disciples of the clan were proud and invincible. In order to break through the cultivation level and ascend as soon as possible, Xu Yang, the only disciple of the sect, had been in the Qi refining stage. When he came out, the cultivation world had already declined and only three or five disciples were left in Tian Lan Clan. As the clan expands, the truth of Xu Yang's stagnant cultivation is revealed step by step, and a hidden secret that runs through the three realms of human, devil and immortal is revealed to everyone! Is it a thought to become a god, or a thought to become a devil? The life and death of the world is between the palms of Xu Yang's hands! (DeepL)",
-        "popularity": 45.9127,
-        "poster_path": "/dZF7DTXgyoyshdALJLOrQ9Zj4Xz.jpg",
-        "first_air_date": "2023-02-18",
-        "name": "One Hundred Thousand Years of Qi Refining",
-        "vote_average": 8.6,
-        "vote_count": 9
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/5jbuPMtZrilFg70CGx47dztBCKf.jpg",
-        "genre_ids": [
-          10763
-        ],
-        "id": 58287,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "PBS News Weekend",
-        "overview": "A summary of the day’s national and international news, using renowned experts to provide in-depth analysis. Each weekend broadcast contains original, in-depth field reporting on topics including education, healthcare, the economy, energy, science and technology, religion, finance and the arts.",
-        "popularity": 46.7905,
-        "poster_path": "/qIccOtBz8uzAbK0HsiB4o0JjpIH.jpg",
-        "first_air_date": "2013-09-07",
-        "name": "PBS News Weekend",
-        "vote_average": 4.8,
-        "vote_count": 12
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/5MWFjQ3NuNtNeOE9FdrVLeNMNPs.jpg",
-        "genre_ids": [
-          16,
-          35,
-          10751,
-          10765
-        ],
-        "id": 1877,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Phineas and Ferb",
-        "overview": "Each day, two kindhearted suburban stepbrothers on summer vacation embark on some grand new project, which annoys their controlling sister, Candace, who tries to bust them. Meanwhile, their pet platypus plots against evil Dr. Doofenshmirtz.",
-        "popularity": 50.9464,
-        "poster_path": "/g6TidT7be954rrzPPBTIXkSmfXj.jpg",
-        "first_air_date": "2007-08-17",
-        "name": "Phineas and Ferb",
-        "vote_average": 7.892,
-        "vote_count": 932
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/ffvaWDSYLcIQRCLFLG1a3lC0xPX.jpg",
-        "genre_ids": [
-          16,
-          35,
-          9648,
-          18
-        ],
-        "id": 82739,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "青春ブタ野郎はバニーガール先輩の夢を見ない",
-        "overview": "Puberty Syndrome—a rumored, mysterious syndrome that only affects those in their puberty. For example, a bunny girl suddenly appeared in front of Sakuta Azusagawa. The bunny girl's real identity is Mai Sakurajima, a teenage celebrity who is currently an inactive high school senior. For some reason, her charming figure does not reflect in the eyes of others. In the course of revealing the mystery behind this phenomenon, Sakuta begins to explore his feelings towards Mai. Set in a city where the skies and seas shine, Sakuta unfolds the meaning behind his bizarre encounters on women with the said syndrome.",
-        "popularity": 46.6898,
-        "poster_path": "/kVFszsNX8f7LnajQSNJqlTOdy0y.jpg",
-        "first_air_date": "2018-10-04",
-        "name": "Rascal Does Not Dream of Bunny Girl Senpai",
-        "vote_average": 8.6,
-        "vote_count": 1251
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/uMVRiXpeqQGu6xKSI4WyNUEke50.jpg",
-        "genre_ids": [
-          10759,
-          18,
-          35
-        ],
-        "id": 121651,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "スターダム・ワールドのスターダム",
-        "overview": "Stardom World is the VOD network for World Wonder Ring Stardom. WWR Stardom is a Japanese women's promotion started in 2011 featuring matches of many varieties including high speed octane action, long brutal brawls, and even costume battle royals. In 2019, the promotion was acquired by Bushiroad and has since become one of the top wrestling promotions in Japan.",
-        "popularity": 42.8885,
-        "poster_path": "/2ZcRPQKAqLghBw0TC0fzMXCnY2o.jpg",
-        "first_air_date": "2011-01-23",
-        "name": "Stardom on Stardom World",
-        "vote_average": 9.5,
-        "vote_count": 2
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/qX9epDSi7n4NGHx9To2HrNPNYty.jpg",
-        "genre_ids": [
-          18,
-          10766
-        ],
-        "id": 233519,
-        "origin_country": [
-          "BR"
-        ],
-        "original_language": "pt",
-        "original_name": "Vale Tudo",
-        "overview": "The story of a mother and daughter with opposing views on the path to success in life takes the audience on a contemporary journey about character, ethics and honesty.",
-        "popularity": 43.0534,
-        "poster_path": "/FH6lI9j3ixwRB1MbToeUTtUzc.jpg",
-        "first_air_date": "2025-03-31",
-        "name": "Vale Tudo",
-        "vote_average": 7,
-        "vote_count": 9
-      }
-    ],
-    "total_pages": 11,
-    "total_results": 206
-  }
-  """#
