--- Tests/TMDBTests/TVShows/Data+OnTheAirTVShows.swift
@@ -1,439 +1,0 @@
-import Foundation
-
-extension Data {
-  static let onTheAirTVShows = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "page": 1,
-    "results": [
-      {
-        "adult": false,
-        "backdrop_path": "/sKlF9YrVu84DYMDAUZEZDCvDxK2.jpg",
-        "genre_ids": [
-          16,
-          35,
-          18,
-          10765
-        ],
-        "id": 244808,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "ぬきたし THE ANIMATION",
-        "overview": "On Seiran Island, secret society NLNS is united by Junnosuke Tachibana, who is determined to crush a law. They have found a benefactor, an unknown rich man, who offers them a secret base and cash. Their mission is to locate a mysterious girl and stop the law, keeping the SHO off their trail.",
-        "popularity": 403.4836,
-        "poster_path": "/zlybZJXfu4IpPnWXhnRtobl9BgZ.jpg",
-        "first_air_date": "2025-07-19",
-        "name": "Nukitashi the Animation",
-        "vote_average": 6.542,
-        "vote_count": 12
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/vgeDRVpSUa4Hvovg4C6dgm4dfUW.jpg",
-        "genre_ids": [
-          10766,
-          18
-        ],
-        "id": 112470,
-        "origin_country": [
-          "FR"
-        ],
-        "original_language": "fr",
-        "original_name": "Ici tout commence",
-        "overview": "This television drama series is centered around the prestigious culinary school of renowned chef Auguste Armand. The show follows the lives of students and staff as they navigate the challenges and pressures of the culinary world—delving into their personal and professional lives, revealing secrets, rivalries, and complex relationships.",
-        "popularity": 359.6653,
-        "poster_path": "/x9HeaagUAyyGl1fQ6exQcpELBxP.jpg",
-        "first_air_date": "2020-11-02",
-        "name": "Ici tout commence",
-        "vote_average": 6.797,
-        "vote_count": 91
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/kCd0rmnNwCd3sgupO9DqSi9aHa0.jpg",
-        "genre_ids": [
-          10767
-        ],
-        "id": 65270,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "라디오스타",
-        "overview": "“Radio Star” is a South Korean music talk show where the guests and hosts talk openly and intimately about their lives, their pasts and their work. It is part of Munhwa Broadcasting Corporation's \"Golden Fishery\" programming block.",
-        "popularity": 353.3722,
-        "poster_path": "/uRUZDsvUfIP3JUEgOC8ReBlQQUU.jpg",
-        "first_air_date": "2007-05-30",
-        "name": "Radio Star",
-        "vote_average": 5.75,
-        "vote_count": 4
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/a8a8z4aJr40zYgRMfmSESppX2eO.jpg",
-        "genre_ids": [
-          10764,
-          35
-        ],
-        "id": 70672,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "아는 형님",
-        "overview": "Male celebs play make-believe as high schoolers, welcoming star transfer students every week and engaging in battles of witty humor and slapstick.",
-        "popularity": 321.8921,
-        "poster_path": "/2jIi55JtYKJTL1km8qHMuUilOWo.jpg",
-        "first_air_date": "2015-12-05",
-        "name": "Men on a Mission",
-        "vote_average": 7.528,
-        "vote_count": 36
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/hINekSpbcBxjnjGqmIm6I4bz2ab.jpg",
-        "genre_ids": [
-          10767,
-          35
-        ],
-        "id": 22980,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Watch What Happens Live with Andy Cohen",
-        "overview": "Bravo network executive Andy Cohen discusses pop culture topics with celebrities and reality show personalities.",
-        "popularity": 312.5306,
-        "poster_path": "/onSD9UXfJwrMXWhq7UY7hGF2S1h.jpg",
-        "first_air_date": "2009-07-16",
-        "name": "Watch What Happens Live with Andy Cohen",
-        "vote_average": 5.1,
-        "vote_count": 77
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/7VO04TtL1jIT6XOPs9u4jdB8KaB.jpg",
-        "genre_ids": [
-          35,
-          10767
-        ],
-        "id": 59941,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "The Tonight Show Starring Jimmy Fallon",
-        "overview": "After Jay Leno's second retirement from the program, Jimmy Fallon stepped in as his permanent replacement. After 42 years in Los Angeles the program was brought back to New York.",
-        "popularity": 277.4346,
-        "poster_path": "/if7ECoH4xaYw5S8gomNIEmtwTxP.jpg",
-        "first_air_date": "2014-02-17",
-        "name": "The Tonight Show Starring Jimmy Fallon",
-        "vote_average": 5.8,
-        "vote_count": 348
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/gMMnf8VRg3Z98WaFmOLr9Jk8pIs.jpg",
-        "genre_ids": [
-          35,
-          10767
-        ],
-        "id": 63770,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "The Late Show with Stephen Colbert",
-        "overview": "Stephen Colbert brings his signature satire and comedy to The Late Show with Stephen Colbert, the #1 show in late night, where he talks with an eclectic mix of guests about what is new and relevant in the worlds of politics, entertainment, business, music, technology, and more. Featuring bandleader Jon Batiste with his band Stay Human, the Emmy Award-nominated show is broadcast from the historic Ed Sullivan Theater.",
-        "popularity": 254.0311,
-        "poster_path": "/9jkThAGYj2yp8jsS6Nriy5mzKFT.jpg",
-        "first_air_date": "2015-09-08",
-        "name": "The Late Show with Stephen Colbert",
-        "vote_average": 6.281,
-        "vote_count": 319
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/32iRX9xqNH3ErJN6FowxdtgYmiY.jpg",
-        "genre_ids": [
-          35,
-          10764
-        ],
-        "id": 30801,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "1박 2일",
-        "overview": "2 Days & 1 Night is a South Korean reality-variety show with the motto \"real wild road variety.\" Its main concept is to recommend various places of interest that viewers can visit in South Korea.",
-        "popularity": 234.1888,
-        "poster_path": "/qezeCEkIKHQM5iwyVeYBuXmYA2h.jpg",
-        "first_air_date": "2007-08-05",
-        "name": "2 Days and 1 Night",
-        "vote_average": 7.1,
-        "vote_count": 27
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/iEg2LHsFV3vC6LywkBAL3c1n9PP.jpg",
-        "genre_ids": [
-          10767,
-          35
-        ],
-        "id": 61818,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Late Night with Seth Meyers",
-        "overview": "Seth Meyers, who is \"Saturday Night Live’s\" longest serving anchor on the show’s wildly popular \"Weekend Update,\" takes over as host of NBC’s \"Late Night\" — home to A-list celebrity guests, memorable comedy and the best in musical talent. As the Emmy Award-winning head writer for \"SNL,\" Meyers has established a reputation for sharp wit and perfectly timed comedy, and has gained fame for his spot-on jokes and satire. Meyers takes his departure from \"SNL\" to his new post at \"Late Night,\" as Jimmy Fallon moves to \"The Tonight Show\".",
-        "popularity": 229.3017,
-        "poster_path": "/rw0QaegwgKbRoB2CZe2lfewltT9.jpg",
-        "first_air_date": "2014-02-25",
-        "name": "Late Night with Seth Meyers",
-        "vote_average": 5.323,
-        "vote_count": 113
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/phc7wS5j4anfu9EoD5NMNRozoDC.jpg",
-        "genre_ids": [
-          10763
-        ],
-        "id": 651,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "60 Minutes",
-        "overview": "America's popular television News magazine in which an ever changing team of CBS News correspondents contribute segments ranging from hard news coverage to politics to lifestyle and pop culture.",
-        "popularity": 240.2282,
-        "poster_path": "/qDSKxeHoN9pvVChXXc2YZN4KwBJ.jpg",
-        "first_air_date": "1968-09-24",
-        "name": "60 Minutes",
-        "vote_average": 6.7,
-        "vote_count": 75
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/jCQV3rHNbgHKo3N4G6CARbVEjPV.jpg",
-        "genre_ids": [
-          18
-        ],
-        "id": 194766,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "The Summer I Turned Pretty",
-        "overview": "Every summer, Belly and her family head to the Fishers’ beach house in Cousins. Every summer is the same ... until Belly turns sixteen. Relationships will be tested, painful truths will be revealed, and Belly will be forever changed. It’s a summer of first love, first heartbreak and growing up — it's the summer she turns pretty.",
-        "popularity": 206.376,
-        "poster_path": "/s85u6WqkfkVBOrWDMKrefnU1g1n.jpg",
-        "first_air_date": "2022-06-16",
-        "name": "The Summer I Turned Pretty",
-        "vote_average": 8.2,
-        "vote_count": 853
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/tNi2aJPdCKfielGYzIi3IKxStz9.jpg",
-        "genre_ids": [
-          10762,
-          35
-        ],
-        "id": 502,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Sesame Street",
-        "overview": "On a special inner city street, the inhabitants—human and muppet—teach preschoolers basic educational and social concepts using comedy, cartoons, games, and songs.",
-        "popularity": 197.3031,
-        "poster_path": "/14k9BfZ2p4rQBMeJ5crKTfUZVwD.jpg",
-        "first_air_date": "1969-11-10",
-        "name": "Sesame Street",
-        "vote_average": 7.1,
-        "vote_count": 267
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/5546GF3zT3yA25ydCz0KlujjRA3.jpg",
-        "genre_ids": [
-          16,
-          80,
-          9648,
-          35
-        ],
-        "id": 30983,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "名探偵コナン",
-        "overview": "The son of a world famous mystery writer, Jimmy Kudo, has achieved his own notoriety by assisting the local police as a student detective. He has always been able to solve the most difficult of criminal cases using his wits and power of reason.",
-        "popularity": 198.2992,
-        "poster_path": "/1o2o7RELi8rtRrTh8iHYkprpDrn.jpg",
-        "first_air_date": "1996-01-08",
-        "name": "Case Closed",
-        "vote_average": 8,
-        "vote_count": 704
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/kkfqNkGQR5og5sDjJTxTVmI9PW.jpg",
-        "genre_ids": [
-          80,
-          18,
-          10766
-        ],
-        "id": 72879,
-        "origin_country": [
-          "FR"
-        ],
-        "original_language": "fr",
-        "original_name": "Demain nous appartient",
-        "overview": "The story revolves around the people of Sète, France. Their lives are punctuated by family rivalries, romance and scenes from daily life, but also by plots involving police investigations, secrets and betrayals.",
-        "popularity": 170.4706,
-        "poster_path": "/zMWldNZF0wS3L5XkDVFHxYhclcL.jpg",
-        "first_air_date": "2017-07-17",
-        "name": "Tomorrow Is Ours",
-        "vote_average": 6.343,
-        "vote_count": 67
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/f1OoogO23fHdm7NzCsCxjFPoQit.jpg",
-        "genre_ids": [
-          10759,
-          16,
-          35,
-          10762,
-          10765,
-          18
-        ],
-        "id": 80559,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "パズドラ",
-        "overview": "The story is set in modern day Japan following the growth of the protagonist Taiga Akashi, an elementary school kid who wants to be a professional gamer someday.",
-        "popularity": 169.3997,
-        "poster_path": "/oW9rE2Zo38eTBcCcnDl5tZijYHq.jpg",
-        "first_air_date": "2018-04-02",
-        "name": "Puzzle & Dragons",
-        "vote_average": 8,
-        "vote_count": 3
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/vrLDuQ5L45f7vFWpMT17WopdWx0.jpg",
-        "genre_ids": [
-          10764,
-          10767
-        ],
-        "id": 72896,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "미운 우리 새끼",
-        "overview": "How much do you know about your grown up son? In Mom’s Diary, the daily lives of the grown up celebrity sons are closely observed by their mothers. Do you know why your son drinks so much these days? How much do you know about your son’s philosophy in his life style? Heart-warming moments of observational communication between a mother and a son can be seen on Mom’s Diary.",
-        "popularity": 177.5392,
-        "poster_path": "/fgeWRIiigNHLxy3qEzM0HqNcFHT.jpg",
-        "first_air_date": "2016-08-26",
-        "name": "My Little Old Boy",
-        "vote_average": 6.375,
-        "vote_count": 8
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/1eUNe8eP5GnEL6RLgPLSVOiVTvU.jpg",
-        "genre_ids": [
-          10751,
-          10764
-        ],
-        "id": 64369,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "TV 동물농장",
-        "overview": "",
-        "popularity": 150.9426,
-        "poster_path": "/7FZZHXxlujrtq3EUkP78tm5aVkX.jpg",
-        "first_air_date": "2001-05-06",
-        "name": "TV Animal Farm",
-        "vote_average": 7.5,
-        "vote_count": 4
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/3z4zE7WvIyQYadqxDQ7MCpSRbOG.jpg",
-        "genre_ids": [
-          10766
-        ],
-        "id": 249010,
-        "origin_country": [
-          "ZA"
-        ],
-        "original_language": "af",
-        "original_name": "Kelders van Geheime",
-        "overview": "Kelders van Geheime (Cellars of Secrets) is a South African, Afrikaans-language telenovela based on the wheeling's and dealings of the community that lives on Soebatskloof (a wine farm in the Cape). The series focuses on three families: the Abrahams, Syster and Marais households. These families will have to stand together, agree (or agree to disagree), and work together to earn their bread and butter – and a glass of wine.",
-        "popularity": 131.3745,
-        "poster_path": "/r0hqXxOTJ7pTEYQYl7HWrD50ym0.jpg",
-        "first_air_date": "2024-05-06",
-        "name": "Kelders van Geheime",
-        "vote_average": 6,
-        "vote_count": 13
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/kSFEIZS3HplqY1uN9U1GR0aEKlP.jpg",
-        "genre_ids": [
-          10764
-        ],
-        "id": 62481,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Beat Bobby Flay",
-        "overview": "Two talented chefs go head-to-head for the chance to Beat Bobby Flay. To get to Bobby the chefs must first face off against each other, creating a spectacular dish with a secret ingredient of Bobby's choice. Judges Alex Guarnaschelli and Jeff Mauro know Bobby's strengths and his weaknesses. Their goal: Pick the chef who has the skills to take down Bobby Flay in his own arena. The winning chef gets to challenge Bobby with his or her surprise signature dish. If Bobby goes down, the winner can tell the world, \"I beat Bobby Flay!\"",
-        "popularity": 198.1762,
-        "poster_path": "/vY8Bep9K9oFSz0C9gnTVipfBciU.jpg",
-        "first_air_date": "2013-08-24",
-        "name": "Beat Bobby Flay",
-        "vote_average": 6.519,
-        "vote_count": 26
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/6WAinokBmZU9Jt8l74TKI8JNUG7.jpg",
-        "genre_ids": [
-          16,
-          35
-        ],
-        "id": 2190,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "South Park",
-        "overview": "Follow the misadventures of four irreverent grade-schoolers in the quiet, dysfunctional town of South Park, Colorado.",
-        "popularity": 146.0586,
-        "poster_path": "/dqAKFEjPEu7vWkVBkMdhGu9ifxU.jpg",
-        "first_air_date": "1997-08-13",
-        "name": "South Park",
-        "vote_average": 8.4,
-        "vote_count": 4609
-      }
-    ],
-    "total_pages": 52,
-    "total_results": 1033
-  }
-  """#
