--- Tests/TMDBTests/TVShows/Data+PopularTVShows.swift
@@ -1,444 +1,0 @@
-import Foundation
-
-extension Data {
-  static let popularTVShows = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "page": 1,
-    "results": [
-      {
-        "adult": false,
-        "backdrop_path": "/hHMWfDlqKisD5zVcZfz3KEw5U3v.jpg",
-        "genre_ids": [
-          18,
-          10751
-        ],
-        "id": 50821,
-        "origin_country": [
-          "HU"
-        ],
-        "original_language": "hu",
-        "original_name": "Barátok közt",
-        "overview": "Miklós introduces viewers into the world of Money and Charm.",
-        "popularity": 492.0376,
-        "poster_path": "/kBBbSgNchtMvsgD6z1oI1RRluHP.jpg",
-        "first_air_date": "1998-10-26",
-        "name": "Among Friends",
-        "vote_average": 3.528,
-        "vote_count": 18
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/hqqdoXdrp4o8ZqADHVYIXwBkn3Y.jpg",
-        "genre_ids": [
-          10759,
-          16,
-          35,
-          10765,
-          10762,
-          10751
-        ],
-        "id": 65733,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "ドラえもん",
-        "overview": "Robotic cat Doraemon is sent back in time from the 22nd century to protect 10-year-old Noby, a lazy and uncoordinated boy who is destined to have a tragic future. Doraemon can create secret gadgets from a pocket on his stomach, but they usually cause more bad than good because of Noby's propensity to misuse them.",
-        "popularity": 416.4487,
-        "poster_path": "/9ZN1P32SHviL3SV51qLivxycvcx.jpg",
-        "first_air_date": "2005-04-22",
-        "name": "Doraemon",
-        "vote_average": 8.061,
-        "vote_count": 188
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/sKlF9YrVu84DYMDAUZEZDCvDxK2.jpg",
-        "genre_ids": [
-          16,
-          35,
-          18,
-          10765
-        ],
-        "id": 244808,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "ぬきたし THE ANIMATION",
-        "overview": "On Seiran Island, secret society NLNS is united by Junnosuke Tachibana, who is determined to crush a law. They have found a benefactor, an unknown rich man, who offers them a secret base and cash. Their mission is to locate a mysterious girl and stop the law, keeping the SHO off their trail.",
-        "popularity": 403.4836,
-        "poster_path": "/zlybZJXfu4IpPnWXhnRtobl9BgZ.jpg",
-        "first_air_date": "2025-07-19",
-        "name": "Nukitashi the Animation",
-        "vote_average": 6.542,
-        "vote_count": 12
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/vgeDRVpSUa4Hvovg4C6dgm4dfUW.jpg",
-        "genre_ids": [
-          10766,
-          18
-        ],
-        "id": 112470,
-        "origin_country": [
-          "FR"
-        ],
-        "original_language": "fr",
-        "original_name": "Ici tout commence",
-        "overview": "This television drama series is centered around the prestigious culinary school of renowned chef Auguste Armand. The show follows the lives of students and staff as they navigate the challenges and pressures of the culinary world—delving into their personal and professional lives, revealing secrets, rivalries, and complex relationships.",
-        "popularity": 359.6653,
-        "poster_path": "/x9HeaagUAyyGl1fQ6exQcpELBxP.jpg",
-        "first_air_date": "2020-11-02",
-        "name": "Ici tout commence",
-        "vote_average": 6.797,
-        "vote_count": 91
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/kCd0rmnNwCd3sgupO9DqSi9aHa0.jpg",
-        "genre_ids": [
-          10767
-        ],
-        "id": 65270,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "라디오스타",
-        "overview": "“Radio Star” is a South Korean music talk show where the guests and hosts talk openly and intimately about their lives, their pasts and their work. It is part of Munhwa Broadcasting Corporation's \"Golden Fishery\" programming block.",
-        "popularity": 353.3722,
-        "poster_path": "/uRUZDsvUfIP3JUEgOC8ReBlQQUU.jpg",
-        "first_air_date": "2007-05-30",
-        "name": "Radio Star",
-        "vote_average": 5.75,
-        "vote_count": 4
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/mSkisNewxu55Ivar5EubhcIuoqn.jpg",
-        "genre_ids": [
-          10764
-        ],
-        "id": 30957,
-        "origin_country": [
-          "US",
-          "GB"
-        ],
-        "original_language": "en",
-        "original_name": "The Amazing Race",
-        "overview": "This reality competition sees teams embark on a trek around the world to amazing destinations where they must compete in a series of challenges, some mental and some physical. Only when the tasks are completed will they learn of their next location. Teams who are the farthest behind will gradually be eliminated as the contest progresses, with the first team to arrive at the final destination winning the race and the $1 million prize.",
-        "popularity": 335.8333,
-        "poster_path": "/x1jq0atcQg6UkaZpco0kFL6QHuM.jpg",
-        "first_air_date": "2001-09-05",
-        "name": "The Amazing Race",
-        "vote_average": 6.868,
-        "vote_count": 155
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/a8a8z4aJr40zYgRMfmSESppX2eO.jpg",
-        "genre_ids": [
-          10764,
-          35
-        ],
-        "id": 70672,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "아는 형님",
-        "overview": "Male celebs play make-believe as high schoolers, welcoming star transfer students every week and engaging in battles of witty humor and slapstick.",
-        "popularity": 321.8921,
-        "poster_path": "/2jIi55JtYKJTL1km8qHMuUilOWo.jpg",
-        "first_air_date": "2015-12-05",
-        "name": "Men on a Mission",
-        "vote_average": 7.4,
-        "vote_count": 37
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/3q7cCjq2cEcYjuOIpL46n11l1Gs.jpg",
-        "genre_ids": [
-          99
-        ],
-        "id": 1966,
-        "origin_country": [
-          "GB"
-        ],
-        "original_language": "en",
-        "original_name": "Horizon",
-        "overview": "Horizon tells amazing science stories, unravels mysteries and reveals worlds you've never seen before.",
-        "popularity": 313.4187,
-        "poster_path": "/uX8pic7asQBdnwMB9QjPvTsn1Dw.jpg",
-        "first_air_date": "1964-02-04",
-        "name": "Horizon",
-        "vote_average": 7.516,
-        "vote_count": 32
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/hINekSpbcBxjnjGqmIm6I4bz2ab.jpg",
-        "genre_ids": [
-          10767,
-          35
-        ],
-        "id": 22980,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Watch What Happens Live with Andy Cohen",
-        "overview": "Bravo network executive Andy Cohen discusses pop culture topics with celebrities and reality show personalities.",
-        "popularity": 312.5306,
-        "poster_path": "/onSD9UXfJwrMXWhq7UY7hGF2S1h.jpg",
-        "first_air_date": "2009-07-16",
-        "name": "Watch What Happens Live with Andy Cohen",
-        "vote_average": 5.1,
-        "vote_count": 77
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/hvFCS0dMeC3ffiF4uYTKzUAkvYL.jpg",
-        "genre_ids": [
-          35
-        ],
-        "id": 65701,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Good Mythical Morning",
-        "overview": "Two \"Internetainers\" (Rhett & Link) go far out and do the weirdest things, giving you a daily dose of casual comedy every Monday-Friday.",
-        "popularity": 306.3686,
-        "poster_path": "/jMpBQr2aNOFAI6wsC47zsOG6qOh.jpg",
-        "first_air_date": "2012-01-09",
-        "name": "Good Mythical Morning",
-        "vote_average": 6.976,
-        "vote_count": 85
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/2m1Mu0xPj4SikiqkaolTRUcNtWH.jpg",
-        "genre_ids": [
-          80,
-          18,
-          35
-        ],
-        "id": 79744,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "The Rookie",
-        "overview": "Starting over isn't easy, especially for small-town guy John Nolan who, after a life-altering incident, is pursuing his dream of being an LAPD officer. As the force's oldest rookie, he’s met with skepticism from some higher-ups who see him as just a walking midlife crisis.",
-        "popularity": 303.7086,
-        "poster_path": "/bL1mwXDnH5fCxqc4S2n40hoVyoe.jpg",
-        "first_air_date": "2018-10-16",
-        "name": "The Rookie",
-        "vote_average": 8.537,
-        "vote_count": 2589
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/2meX1nMdScFOoV4370rqHWKmXhY.jpg",
-        "genre_ids": [
-          10759,
-          9648,
-          18
-        ],
-        "id": 93405,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "오징어 게임",
-        "overview": "Hundreds of cash-strapped players accept a strange invitation to compete in children's games. Inside, a tempting prize awaits — with deadly high stakes.",
-        "popularity": 301.2045,
-        "poster_path": "/1QdXdRYfktUSONkl1oD5gc6Be0s.jpg",
-        "first_air_date": "2021-09-17",
-        "name": "Squid Game",
-        "vote_average": 7.865,
-        "vote_count": 16489
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/fgvM6m7gAd8kDAaLF16aflbOSAK.jpg",
-        "genre_ids": [
-          18
-        ],
-        "id": 1416,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Grey's Anatomy",
-        "overview": "Follows the personal and professional lives of a group of doctors at Seattle’s Grey Sloan Memorial Hospital.",
-        "popularity": 281.3067,
-        "poster_path": "/7jEVqXC14bhfAzSPgr896dMdDv6.jpg",
-        "first_air_date": "2005-03-27",
-        "name": "Grey's Anatomy",
-        "vote_average": 8.221,
-        "vote_count": 10581
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/7VO04TtL1jIT6XOPs9u4jdB8KaB.jpg",
-        "genre_ids": [
-          35,
-          10767
-        ],
-        "id": 59941,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "The Tonight Show Starring Jimmy Fallon",
-        "overview": "After Jay Leno's second retirement from the program, Jimmy Fallon stepped in as his permanent replacement. After 42 years in Los Angeles the program was brought back to New York.",
-        "popularity": 277.4346,
-        "poster_path": "/if7ECoH4xaYw5S8gomNIEmtwTxP.jpg",
-        "first_air_date": "2014-02-17",
-        "name": "The Tonight Show Starring Jimmy Fallon",
-        "vote_average": 5.834,
-        "vote_count": 349
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/lirPqYLTtd6XZqGn4cS1wiesTq0.jpg",
-        "genre_ids": [
-          18,
-          9648,
-          10765
-        ],
-        "id": 1622,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Supernatural",
-        "overview": "When they were boys, Sam and Dean Winchester lost their mother to a mysterious and demonic supernatural force. Subsequently, their father raised them to be soldiers. He taught them about the paranormal evil that lives in the dark corners and on the back roads of America ... and he taught them how to kill it. Now, the Winchester brothers crisscross the country in their '67 Chevy Impala, battling every kind of supernatural threat they encounter along the way. ",
-        "popularity": 263.0525,
-        "poster_path": "/KoYWXbnYuS3b0GyQPkbuexlVK9.jpg",
-        "first_air_date": "2005-09-13",
-        "name": "Supernatural",
-        "vote_average": 8.3,
-        "vote_count": 7916
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/qLMfcvdCcCGD2BNLH8b6ZCBuO7D.jpg",
-        "genre_ids": [
-          10763,
-          35
-        ],
-        "id": 2224,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "The Daily Show",
-        "overview": "The World's Fakest News Team tackle the biggest stories in news, politics and pop culture.",
-        "popularity": 262.2261,
-        "poster_path": "/ixcfyK7it6FjRM36Te4OdblAq4X.jpg",
-        "first_air_date": "1996-07-22",
-        "name": "The Daily Show",
-        "vote_average": 6.381,
-        "vote_count": 590
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/tJF9M5niAnb4fm881PGc2GwS0e1.jpg",
-        "genre_ids": [
-          16,
-          10759,
-          35,
-          10751,
-          10762
-        ],
-        "id": 80885,
-        "origin_country": [
-          "JP"
-        ],
-        "original_language": "ja",
-        "original_name": "忍者ハットリくん",
-        "overview": "11-year-old Kenichi Mitsuba is an average kid who goes to secondary school and struggles with his studies, he is very stubborn and is very lazy and therefore always ends up frustrating his parents and teacher. He loves to find an easy way of things. He befriends Hattori Kanzo, a ninja from the Iga Clan, and he becomes part of the Mitsuba family along with his brother, Shinzo and his ninja dog, Shishimaru. Hattori helps Kenichi with his problems, and constantly keeps an eye on him, as a good friend. The main antagonist Kemumaki, a Koga ninja and his ninja cat, Kagechiyo always troubles Kenichi, mainly because of their feud over one girl, Yumeko. Kenichi asks Hattori to take revenge as a recurring storyline in many episodes. Although Hattori is a good friend, Kenichi sometimes fights with Hattori due to misunderstandings created by Kemumaki. Sometimes Jippou, Togejirou and Tsubame help him.",
-        "popularity": 258.3693,
-        "poster_path": "/9PUkpvq0wJnJ2M00dtgxKn6LptB.jpg",
-        "first_air_date": "1981-09-28",
-        "name": "Ninja Hattori-kun",
-        "vote_average": 6.7,
-        "vote_count": 31
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/92cc2hINaW72HTA0RqQaqPUbl4g.jpg",
-        "genre_ids": [
-          16,
-          35,
-          10767
-        ],
-        "id": 46825,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "Rooster Teeth Animated Adventures",
-        "overview": "The animated shenanigans of the Rooster Teeth staff. Audio taken from the Rooster Teeth Podcast.",
-        "popularity": 254.7297,
-        "poster_path": "/xrcOsjmuBcmf1YhqyQ6qrGCcsvE.jpg",
-        "first_air_date": "2011-09-28",
-        "name": "Rooster Teeth Animated Adventures",
-        "vote_average": 8.6,
-        "vote_count": 5
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/gMMnf8VRg3Z98WaFmOLr9Jk8pIs.jpg",
-        "genre_ids": [
-          35,
-          10767
-        ],
-        "id": 63770,
-        "origin_country": [
-          "US"
-        ],
-        "original_language": "en",
-        "original_name": "The Late Show with Stephen Colbert",
-        "overview": "Stephen Colbert brings his signature satire and comedy to The Late Show with Stephen Colbert, the #1 show in late night, where he talks with an eclectic mix of guests about what is new and relevant in the worlds of politics, entertainment, business, music, technology, and more. Featuring bandleader Jon Batiste with his band Stay Human, the Emmy Award-nominated show is broadcast from the historic Ed Sullivan Theater.",
-        "popularity": 254.0311,
-        "poster_path": "/9jkThAGYj2yp8jsS6Nriy5mzKFT.jpg",
-        "first_air_date": "2015-09-08",
-        "name": "The Late Show with Stephen Colbert",
-        "vote_average": 6.281,
-        "vote_count": 319
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/s2qJb1FLwp1Mse9lY08JqgMp0kw.jpg",
-        "genre_ids": [
-          10764,
-          35
-        ],
-        "id": 5092,
-        "origin_country": [
-          "KR"
-        ],
-        "original_language": "ko",
-        "original_name": "무한도전",
-        "overview": "Infinite Challenge has been reported as the first \"Real-Variety\" show in Korean television history. The program is largely unscripted, and follows a similar format of challenge-based Reality Television programs, familiar to the audiences in the West, but the challenges are often silly, absurd, or impossible to achieve, so the program takes on the aspect of a satirical comedy variety show, rather than a more standard reality or contest program. In order to achieve its comedic purposes its 6 hosts and staff continuously proclaim, the elements of this show are the 3-Ds, Dirty, Dangerous, and Difficult.",
-        "popularity": 245.8381,
-        "poster_path": "/3ZIPTvMnzI5ThmdGeEYQFxJV5Sg.jpg",
-        "first_air_date": "2005-04-23",
-        "name": "Infinite Challenge",
-        "vote_average": 7.7,
-        "vote_count": 20
-      }
-    ],
-    "total_pages": 10111,
-    "total_results": 202205
-  }
-  """#
