--- Tests/TMDBTests/TVShows/Data+TVShow1.swift
@@ -1,26 +1,0 @@
-import Foundation
-
-extension Data {
-  static let tvShow1 = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "adult": false,
-    "backdrop_path": "/5BlPu7u7fRKNkNf3b3JPCpt249L.jpg",
-    "genre_ids": [],
-    "id": 297359,
-    "origin_country": [
-      "JP"
-    ],
-    "original_language": "en",
-    "original_name": "J-Trip Plan",
-    "overview": "This travel show aimed at overseas visitors offers info on where to go, what to see, and how to experience Japan, like an expert.",
-    "popularity": 0,
-    "poster_path": "/qcDBu6cKO1rAphFDKwTWcPDwgdm.jpg",
-    "first_air_date": "2016-04-04",
-    "name": "J-Trip Plan",
-    "vote_average": 0,
-    "vote_count": 0
-  }
-  """#
