--- Tests/TMDBTests/TVShows/Data+TVShow2.swift
@@ -1,28 +1,0 @@
-import Foundation
-
-extension Data {
-  static let tvShow2 = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "adult": false,
-    "backdrop_path": "/umpv38zP1GMNU9TTJUH6HZP4jby.jpg",
-    "genre_ids": [
-      10764
-    ],
-    "id": 297366,
-    "origin_country": [
-      "SE"
-    ],
-    "original_language": "sv",
-    "original_name": "Gift med fotbollen",
-    "overview": "",
-    "popularity": 0,
-    "poster_path": "/saHo4p3JFNH38VdPIXXJD2cojkZ.jpg",
-    "first_air_date": "2025-07-21",
-    "name": "Gift med fotbollen",
-    "vote_average": 0,
-    "vote_count": 0
-  }
-  """#
