--- Tests/TMDBTests/TVShows/Data+TVShowDetailsComplete.swift
@@ -1,630 +1,0 @@
-import Foundation
-
-extension Data {
-  static let tvShowDetailsComplete = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "adult": false,
-    "backdrop_path": "/mXtL24G2Awyl3lKFtreMHnH5qFC.jpg",
-    "created_by": [
-      {
-        "id": 47915,
-        "credit_id": "5253395519c295794005e12e",
-        "name": "Darren Star",
-        "original_name": "Darren Star",
-        "gender": 2,
-        "profile_path": "/lh6yexxHGKpLViBUBCWvBzP8kIq.jpg"
-      }
-    ],
-    "episode_run_time": [],
-    "first_air_date": "1998-06-06",
-    "genres": [
-      {
-        "id": 35,
-        "name": "Comedy"
-      },
-      {
-        "id": 18,
-        "name": "Drama"
-      }
-    ],
-    "homepage": "http://www.hbo.com/sex-and-the-city",
-    "id": 105,
-    "in_production": false,
-    "languages": [
-      "en"
-    ],
-    "last_air_date": "2004-02-22",
-    "last_episode_to_air": {
-      "id": 1011398,
-      "name": "An American Girl in Paris (Part Deux)",
-      "overview": "Carrie is desperately lonely in Paris as Petrovsky is spending so much time on his new exhibit. The chemo treatments are killing Samantha's sex drive. Miranda must take care of Steve's mother after she has a stroke. Charlotte and Harry try to adopt a baby. Big finds Carrie in Paris and says the words she's been waiting six years to hear.",
-      "vote_average": 6.2,
-      "vote_count": 5,
-      "air_date": "2004-02-22",
-      "episode_number": 20,
-      "episode_type": "finale",
-      "production_code": "620",
-      "runtime": 43,
-      "season_number": 6,
-      "show_id": 105,
-      "still_path": "/n3ue82ZcCMSKsSUnpRlAAFlyqxn.jpg"
-    },
-    "name": "Sex and the City",
-    "next_episode_to_air": null,
-    "networks": [
-      {
-        "id": 49,
-        "logo_path": "/tuomPhY2UtuPTqqFnKMVHvSb724.png",
-        "name": "HBO",
-        "origin_country": "US"
-      }
-    ],
-    "number_of_episodes": 94,
-    "number_of_seasons": 6,
-    "origin_country": [
-      "US"
-    ],
-    "original_language": "en",
-    "original_name": "Sex and the City",
-    "overview": "Based on the bestselling book by Candace Bushnell, Sex and the City tells the story of four best friends, all single and in their late thirties, as they pursue their careers and talk about their sex lives, all while trying to survive the New York social scene. ",
-    "popularity": 51.1913,
-    "poster_path": "/jfLp8gTfdi9d8onEFJ60kp1Bl1e.jpg",
-    "production_companies": [
-      {
-        "id": 1694,
-        "logo_path": null,
-        "name": "Darren Star Productions",
-        "origin_country": "US"
-      },
-      {
-        "id": 3268,
-        "logo_path": "/tuomPhY2UtuPTqqFnKMVHvSb724.png",
-        "name": "HBO",
-        "origin_country": "US"
-      }
-    ],
-    "production_countries": [
-      {
-        "iso_3166_1": "US",
-        "name": "United States of America"
-      }
-    ],
-    "seasons": [
-      {
-        "air_date": "2003-10-01",
-        "episode_count": 4,
-        "id": 105071,
-        "name": "Specials",
-        "overview": "",
-        "poster_path": "/s3JtxAtad3z2rzyDj2PAsGJ0KW3.jpg",
-        "season_number": 0,
-        "vote_average": 0
-      },
-      {
-        "air_date": "1998-06-06",
-        "episode_count": 12,
-        "id": 60249,
-        "name": "Season 1",
-        "overview": "Thirtysomething writer Carrie Bradshaw's experiences on the New York singles scene serve as the inspiration for her newspaper column: 'Sex and the City'. The struggles of finding a partner are seen through the eyes of Carrie and her best friends Samantha, Charlotte, and Miranda.",
-        "poster_path": "/cPxf5Y1aRE4iq5j08vd1y13cZMK.jpg",
-        "season_number": 1,
-        "vote_average": 5.6
-      },
-      {
-        "air_date": "1999-06-06",
-        "episode_count": 18,
-        "id": 60250,
-        "name": "Season 2",
-        "overview": "Carrie begins dating again in the wake of her break-up with Mr. Big while Samantha, Charlotte and Miranda continue to provide fodder for her column.",
-        "poster_path": "/1ZXJv019e5l2V0DWhD1IFGOGpKC.jpg",
-        "season_number": 2,
-        "vote_average": 6.3
-      },
-      {
-        "air_date": "2000-06-04",
-        "episode_count": 18,
-        "id": 60251,
-        "name": "Season 3",
-        "overview": "",
-        "poster_path": "/24JqcKtnduYEmsbTQo9upDM6cNz.jpg",
-        "season_number": 3,
-        "vote_average": 6.2
-      },
-      {
-        "air_date": "2001-06-03",
-        "episode_count": 18,
-        "id": 60252,
-        "name": "Season 4",
-        "overview": "",
-        "poster_path": "/ive3mic73CQqTGb5Hh9hLGNovfm.jpg",
-        "season_number": 4,
-        "vote_average": 6.1
-      },
-      {
-        "air_date": "2002-07-21",
-        "episode_count": 8,
-        "id": 60253,
-        "name": "Season 5",
-        "overview": "",
-        "poster_path": "/ixtl8ThPcDL7vd310uoHDdkt6vc.jpg",
-        "season_number": 5,
-        "vote_average": 6.4
-      },
-      {
-        "air_date": "2003-06-22",
-        "episode_count": 20,
-        "id": 60254,
-        "name": "Season 6",
-        "overview": "",
-        "poster_path": "/nYyY7AICJFCOhozkacJrHcYGbme.jpg",
-        "season_number": 6,
-        "vote_average": 5.7
-      }
-    ],
-    "spoken_languages": [
-      {
-        "english_name": "English",
-        "iso_639_1": "en",
-        "name": "English"
-      }
-    ],
-    "status": "Ended",
-    "tagline": "The next best thing to being there.",
-    "type": "Scripted",
-    "vote_average": 7.387,
-    "vote_count": 2070,
-    "keywords": {
-      "results": [
-        {
-          "name": "new york city",
-          "id": 242
-        },
-        {
-          "name": "beauty",
-          "id": 3258
-        },
-        {
-          "name": "female friendship",
-          "id": 5248
-        },
-        {
-          "name": "love",
-          "id": 9673
-        },
-        {
-          "name": "dating",
-          "id": 13043
-        },
-        {
-          "name": "newspaper reporter",
-          "id": 163098
-        },
-        {
-          "name": "manhattan, new york city",
-          "id": 168346
-        },
-        {
-          "name": "sitcom",
-          "id": 193171
-        }
-      ]
-    },
-    "similar": {
-      "page": 1,
-      "results": [
-        {
-          "adult": false,
-          "backdrop_path": null,
-          "genre_ids": [
-            35
-          ],
-          "id": 4263,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "The New Dick Van Dyke Show",
-          "overview": "The New Dick Van Dyke Show is an American sitcom starring Dick Van Dyke that aired on CBS from 1971 to 1974. It was Van Dyke's first return to series television since The Dick Van Dyke Show.",
-          "popularity": 12.0966,
-          "poster_path": "/enC1neRtm1mGHnQtMIAvwhCnKO0.jpg",
-          "first_air_date": "1971-09-18",
-          "name": "The New Dick Van Dyke Show",
-          "vote_average": 4,
-          "vote_count": 3
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/azDlPDaWs0d7yxVA4Kr3ka8l2kE.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4277,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Archie Bunker's Place",
-          "overview": "Archie Bunker's Place is an American sitcom originally broadcast on the CBS network, conceived in 1979 as a spin-off and continuation of All in the Family. While not as popular as its predecessor, the show maintained a large enough audience to last for four seasons, until its cancellation in 1983. In its first season, the show performed so well that it knocked Mork & Mindy out of its new Sunday night time slot.",
-          "popularity": 22.9298,
-          "poster_path": "/kjNp6j9XIfUtNrIJWOC9uzfvZN4.jpg",
-          "first_air_date": "1979-09-23",
-          "name": "Archie Bunker's Place",
-          "vote_average": 6.5,
-          "vote_count": 20
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/hS2xAK3X1oWiaA8OX5Bow2SF6KN.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4292,
-          "origin_country": [
-            "GB"
-          ],
-          "original_language": "en",
-          "original_name": "Bottom",
-          "overview": "Richie Richard (socially awkward, sexually inexperienced) and Eddie Hitler (carefree alcoholic ) are two social outcasts living on the dole. Trapped together in a squalid flat in Hammersmith, London they are perpetually skint, bored and sexually frustrated. They spend their days scheming, bickering, and being nasty and sadistic to each other.",
-          "popularity": 7.5291,
-          "poster_path": "/bOrajiWj3Rax8KZ4ZZWcw2KpJD6.jpg",
-          "first_air_date": "1991-09-17",
-          "name": "Bottom",
-          "vote_average": 7.9,
-          "vote_count": 128
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/izWGnAJpD6Lthmq39s7Uyja2RgW.jpg",
-          "genre_ids": [
-            99,
-            10759,
-            80,
-            18
-          ],
-          "id": 4297,
-          "origin_country": [
-            "CA",
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "The Equalizer",
-          "overview": "Robert McCall is a former agent of a secret government agency who is now running his own private crime fighting operation where he fashions himself as \"The Equalizer.\" It is a service for victims of the system who have exhausted all possible means of seeking justice and have nowhere to go. McCall promises to even out the odds for them.",
-          "popularity": 33.2247,
-          "poster_path": "/7nrKPxGJWYe5cacNN9AV9bVplNZ.jpg",
-          "first_air_date": "1985-09-18",
-          "name": "The Equalizer",
-          "vote_average": 7.197,
-          "vote_count": 61
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/jrOUuhskiGNiHbEcAQv8xsh4uPF.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4298,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Campus Ladies",
-          "overview": "Campus Ladies is an American sitcom that premiered on Oxygen on January 8, 2006. It stars Christen Sussin and Carrie Aizley as two 40-something women who decide to go enroll at the fictional University of the Midwest. The humor comes from the various situations the \"ladies\" get involved in while trying to fit in with their much younger friends and classmates.",
-          "popularity": 8.6864,
-          "poster_path": "/nxvJawrY41TIWT9hCLHsM56Syyf.jpg",
-          "first_air_date": "2006-01-08",
-          "name": "Campus Ladies",
-          "vote_average": 3.5,
-          "vote_count": 2
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/qpeuIDFowSiZhNsdcyHkvKPeI1G.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4306,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Oliver Beene",
-          "overview": "Oliver Beene is an American sitcom. Set in 1962 and 1963, the show chronicled the trials and tribulations of the 11-to-12-year-old Oliver Beene, in first person perspective. Oliver Beene's other main characters are his parents Jerry and Charlotte Beene, his brother Ted Beene, and his two friends Joyce and Michael. The narrator, an older Oliver reflecting on his experience, is voiced by David Cross. Often in episodes, the story is interrupted by flashbacks and flash-forwards.",
-          "popularity": 10.5202,
-          "poster_path": "/5dTw5Nawhl5gv8cxj8ESEhGUamW.jpg",
-          "first_air_date": "2003-03-09",
-          "name": "Oliver Beene",
-          "vote_average": 8,
-          "vote_count": 8
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/jgJiZrs7pHgQ8x7BJQbM1PLzPQL.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4311,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Stacked",
-          "overview": "Skyler Dayton trades the beach for a family-run independent bookstore where work rivalries and comic foibles revolve around her bombshell arrival.",
-          "popularity": 7.9642,
-          "poster_path": "/2I8vnE3HzjpEf93pS54C2yvPOYf.jpg",
-          "first_air_date": "2005-04-13",
-          "name": "Stacked",
-          "vote_average": 7.2,
-          "vote_count": 16
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/btUwsz9ttVBVyqtEyZvFBQFyJgv.jpg",
-          "genre_ids": [
-            35,
-            10751
-          ],
-          "id": 4313,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Full House",
-          "overview": "After the death of his wife, Danny enlists his best friend and his brother-in-law to help raise his three daughters, D.J., Stephanie, and Michelle.",
-          "popularity": 35.7233,
-          "poster_path": "/7g0EyKsIaYjYw3gCIBKMHHE0Kcu.jpg",
-          "first_air_date": "1987-09-22",
-          "name": "Full House",
-          "vote_average": 7.54,
-          "vote_count": 1058
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/3ILUdTEetxZZ8RkPxf5e0H9pCcP.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4327,
-          "origin_country": [
-            "GB"
-          ],
-          "original_language": "en",
-          "original_name": "Mr. Bean",
-          "overview": "Mr Bean turns simple everyday tasks into chaotic situations and will leave you in stitches as he creates havoc wherever he goes.",
-          "popularity": 23.3209,
-          "poster_path": "/pUNQV35kDaKuejnXGNiUeEBurNE.jpg",
-          "first_air_date": "1990-01-01",
-          "name": "Mr. Bean",
-          "vote_average": 7.952,
-          "vote_count": 1426
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/lTVQIpIQbh7Rrwuhv8fBmM7BHaJ.jpg",
-          "genre_ids": [
-            18
-          ],
-          "id": 4344,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Rescue Me",
-          "overview": "Rescue Me revolves around the lives of the men in a New York City firehouse, the crew of 62 Truck. Examining the fraternal nature and relationships of firefighters, the series tackles the daily drama of the life-and-death situations associated with being a firefighter, while exploring the ways the men use dark humor to protect their true emotions.",
-          "popularity": 8.1116,
-          "poster_path": "/gnOwOb82chSJkQEZ9pd6jf2ZrEX.jpg",
-          "first_air_date": "2004-07-21",
-          "name": "Rescue Me",
-          "vote_average": 7.642,
-          "vote_count": 172
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/d0nAuGAH5EIQDLTs7dJcQOubpha.jpg",
-          "genre_ids": [
-            10751,
-            35
-          ],
-          "id": 4345,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Saved by the Bell",
-          "overview": "Lovable schemer Zack Morris leads his pals on adventures at California's Bayside High School. The friends navigate relationships, final exams, school dances, breakups and more while frequently frustrating their principal, Mr. Richard Belding, who does his best to keep them in check.",
-          "popularity": 23.5965,
-          "poster_path": "/9YC576fs6XQR2pfPDJn4kcmG8T5.jpg",
-          "first_air_date": "1989-08-20",
-          "name": "Saved by the Bell",
-          "vote_average": 7.81,
-          "vote_count": 670
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/aGo9zYMFDHTVA9dYGhdlHNJgNKt.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4347,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "All-American Girl",
-          "overview": "All-American Girl is a 1994 ABC situation comedy starring Margaret Cho and featuring Jodi Long, Clyde Kusatsu, Amy Hill, B.D. Wong, and J.B. Quon as her Korean-American family.\n\nIt is the second American sitcom centered on a person of Asian descent, namely Cho.\n\nNotable guest stars during the run of the show include Oprah Winfrey, Jack Black, David Cross, Ming-Na, Vicki Lawrence, Quentin Tarantino, Tsai Chin, Mariska Hargitay, Billy Burke, Robert Clohessy and Garrett Wang.\n\nDiedrich Bader was a one time regular in the last episode of All American Girl, which was a pseudo pilot for a proposed but unrealized version of All American Girl, before achieving fame on The Drew Carey Show. On the DVD commentary for the series, Margaret Cho revealed that most of All American Girl's set furniture was reused by The Drew Carey Show.",
-          "popularity": 7.515,
-          "poster_path": "/kSk299r1Ug0XqVfOSN9V9JuJwGQ.jpg",
-          "first_air_date": "1994-09-14",
-          "name": "All-American Girl",
-          "vote_average": 6,
-          "vote_count": 2
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/inMT72sx7shOp5GtXRYmZsbGieD.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4362,
-          "origin_country": [
-            "GB"
-          ],
-          "original_language": "en",
-          "original_name": "Solo",
-          "overview": "Gemma Palmer is 30 years old, and fed up with being betrayed and taken for granted. She's just found out her boyfriend, Danny, has been having sex with her friend, Gloria, and this is the last straw. She throws Danny out, tells Gloria to go jump in a lake, tells off all the people who have been treating her like a doormat, and quits her boring job at an estate agent's. From now on, she's going solo. At least, that is her firm intention.",
-          "popularity": 3.2584,
-          "poster_path": "/wrOvbWQ0V9eU18xJTEOrFrxyMlT.jpg",
-          "first_air_date": "1981-01-11",
-          "name": "Solo",
-          "vote_average": 6.8,
-          "vote_count": 6
-        },
-        {
-          "adult": false,
-          "backdrop_path": null,
-          "genre_ids": [
-            35
-          ],
-          "id": 4363,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Baby, I'm Back",
-          "overview": "Baby, I'm Back is an American sitcom that aired CBS from January to April 1978. The series stars former Sanford and Son star Demond Wilson, Room 222 alumna Denise Nicholas, Helen Martin, and future Facts of Life co-star Kim Fields.",
-          "popularity": 5.6965,
-          "poster_path": "/eSbABxSoBwo3xti3UatVAbf12pm.jpg",
-          "first_air_date": "1978-01-30",
-          "name": "Baby, I'm Back",
-          "vote_average": 5,
-          "vote_count": 1
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/h7qobus8Ic3vBqF5kefMP7o0jrh.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4388,
-          "origin_country": [
-            "GB"
-          ],
-          "original_language": "en",
-          "original_name": "The Royle Family",
-          "overview": "British comedy series focusing on the lives of a working-class family in Manchester who love the TV.",
-          "popularity": 10.8494,
-          "poster_path": "/bx0JJyMiMnYJdCtryxQCrRe2aen.jpg",
-          "first_air_date": "1998-09-14",
-          "name": "The Royle Family",
-          "vote_average": 7.6,
-          "vote_count": 54
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/y3p2SxZw5hRHO906UckifLbRDjL.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 4402,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "The New Adventures of Old Christine",
-          "overview": "Single working mom Christine Campbell has just learned that her ex is dating a much younger woman with the same first name. To avoid any confusion, the new girlfriend is dubbed New Christine, which leaves her with the unfortunate nickname Old Christine.",
-          "popularity": 27.1386,
-          "poster_path": "/j1u9EGyEQEVFjl5VxK7NbRCxdZ1.jpg",
-          "first_air_date": "2006-03-13",
-          "name": "The New Adventures of Old Christine",
-          "vote_average": 7.032,
-          "vote_count": 95
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/nV9DvSaMcjGNr5eEWxoiIFmMPtC.jpg",
-          "genre_ids": [
-            18,
-            35
-          ],
-          "id": 4412,
-          "origin_country": [
-            "US"
-          ],
-          "original_language": "en",
-          "original_name": "Love Monkey",
-          "overview": "Thirty-something, single, record executive Tom Ferrell navigates the tumultuous and highly amusing waters of work and dating in New York City.",
-          "popularity": 6.5752,
-          "poster_path": "/v7LxXX1mEK26dAJP15ASyUmIynb.jpg",
-          "first_air_date": "2006-01-17",
-          "name": "Love Monkey",
-          "vote_average": 7.5,
-          "vote_count": 8
-        },
-        {
-          "adult": false,
-          "backdrop_path": "/f58UXFInoc2GLHTZHoxbqRzH4H6.jpg",
-          "genre_ids": [
-            35
-          ],
-          "id": 275138,
-          "origin_country": [
-            "GB"
-          ],
-          "original_language": "en",
-          "original_name": "Zipwaz Humphries",
-          "overview": "Surreal sitcom about a kindly idiot who lives with an insecure fireman.",
-          "popularity": 1.8793,
-          "poster_path": "/ur3iQ03x2mRCzwX1JEqWtZoMFq7.jpg",
-          "first_air_date": "2015-06-01",
-          "name": "Zipwaz Humphries",
-          "vote_average": 0,
-          "vote_count": 0
-        },
-        {
-          "adult": false,
-          "backdrop_path": null,
-          "genre_ids": [
-            35
-          ],
-          "id": 275202,
-          "origin_country": [
-            "MX"
-          ],
-          "original_language": "es",
-          "original_name": "La oficina",
-          "overview": "Regional manager Jerónimo Ponce III leads the Aguascalientes branch of the family-run Jabones Olimpo, where the employees have to endure the headaches of working for a family business where the boss is not qualified to lead the company.",
-          "popularity": 0.3598,
-          "poster_path": null,
-          "first_air_date": "",
-          "name": "La oficina",
-          "vote_average": 0,
-          "vote_count": 0
-        },
-        {
-          "adult": false,
-          "backdrop_path": null,
-          "genre_ids": [
-            18
-          ],
-          "id": 122403,
-          "origin_country": [
-            "CN"
-          ],
-          "original_language": "zh",
-          "original_name": "东西奇遇结良缘",
-          "overview": "",
-          "popularity": 1.0657,
-          "poster_path": "/oP71yVsH4fKbQ4uW7kt7wGkh2sG.jpg",
-          "first_air_date": "2001-08-24",
-          "name": "Dong Xi Qi Yu Jie Liang Yuan",
-          "vote_average": 0,
-          "vote_count": 0
-        }
-      ],
-      "total_pages": 3278,
-      "total_results": 65543
-    }
-  }
-  """#
