--- Tests/TMDBTests/TVShows/Data+TVShowDetailsMinimal.swift
@@ -1,57 +1,0 @@
-import Foundation
-
-extension Data {
-  static let tvShowDetailsMinimal = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "adult": false,
-    "backdrop_path": null,
-    "created_by": [],
-    "episode_run_time": [],
-    "first_air_date": "",
-    "genres": [],
-    "homepage": "",
-    "id": 297377,
-    "in_production": true,
-    "languages": [],
-    "last_air_date": null,
-    "last_episode_to_air": {
-      "id": 6425387,
-      "name": "Episode 1",
-      "overview": "",
-      "vote_average": 0,
-      "vote_count": 0,
-      "air_date": "2025-02-07",
-      "episode_number": 1,
-      "episode_type": "standard",
-      "production_code": "",
-      "runtime": null,
-      "season_number": 1,
-      "still_path": null
-    },
-    "name": "仙台有树（重剪版）",
-    "next_episode_to_air": null,
-    "networks": [],
-    "number_of_episodes": null,
-    "number_of_seasons": 0,
-    "origin_country": [
-      "CN"
-    ],
-    "original_language": "zh",
-    "original_name": "仙台有树（重剪版）",
-    "overview": "",
-    "popularity": 0,
-    "poster_path": null,
-    "production_companies": [],
-    "production_countries": [],
-    "seasons": [],
-    "spoken_languages": [],
-    "status": "Returning Series",
-    "tagline": "",
-    "type": "Scripted",
-    "vote_average": 0,
-    "vote_count": 0
-  }
-  """#
