+++ Tests/TMDBTests/TVShows/Data+TVShowDetailsMock.swift
@@ -1,0 +1,334 @@
+//
+//  Data+TVShowDetailsMock.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+import Foundation
+
+extension Data {
+  static let tvShowDetailsComplete = Data(content.utf8)
+}
+
+private let content = #"""
+  {
+    "adult": false,
+    "backdrop_path": "/mXtL24G2Awyl3lKFtreMHnH5qFC.jpg",
+    "created_by": [
+      {
+        "id": 47915,
+        "credit_id": "5253395519c295794005e12e",
+        "name": "Darren Star",
+        "original_name": "Darren Star",
+        "gender": 2,
+        "profile_path": "/lh6yexxHGKpLViBUBCWvBzP8kIq.jpg"
+      }
+    ],
+    "episode_run_time": [],
+    "first_air_date": "1998-06-06",
+    "genres": [
+      {
+        "id": 35,
+        "name": "Comedy"
+      },
+      {
+        "id": 18,
+        "name": "Drama"
+      }
+    ],
+    "homepage": "http://www.hbo.com/sex-and-the-city",
+    "id": 105,
+    "in_production": false,
+    "languages": [
+      "en"
+    ],
+    "last_air_date": "2004-02-22",
+    "last_episode_to_air": {
+      "id": 1011398,
+      "name": "An American Girl in Paris (Part Deux)",
+      "overview": "Carrie is desperately lonely in Paris as Petrovsky is spending so much time on his new exhibit. The chemo treatments are killing Samantha's sex drive. Miranda must take care of Steve's mother after she has a stroke. Charlotte and Harry try to adopt a baby. Big finds Carrie in Paris and says the words she's been waiting six years to hear.",
+      "vote_average": 6.2,
+      "vote_count": 5,
+      "air_date": "2004-02-22",
+      "episode_number": 20,
+      "episode_type": "finale",
+      "production_code": "620",
+      "runtime": 43,
+      "season_number": 6,
+      "show_id": 105,
+      "still_path": "/n3ue82ZcCMSKsSUnpRlAAFlyqxn.jpg"
+    },
+    "name": "Sex and the City",
+    "next_episode_to_air": null,
+    "networks": [
+      {
+        "id": 49,
+        "logo_path": "/tuomPhY2UtuPTqqFnKMVHvSb724.png",
+        "name": "HBO",
+        "origin_country": "US"
+      }
+    ],
+    "number_of_episodes": 94,
+    "number_of_seasons": 6,
+    "origin_country": [
+      "US"
+    ],
+    "original_language": "en",
+    "original_name": "Sex and the City",
+    "overview": "Based on the bestselling book by Candace Bushnell, Sex and the City tells the story of four best friends, all single and in their late thirties, as they pursue their careers and talk about their sex lives, all while trying to survive the New York social scene. ",
+    "popularity": 51.1913,
+    "poster_path": "/jfLp8gTfdi9d8onEFJ60kp1Bl1e.jpg",
+    "production_companies": [
+      {
+        "id": 1694,
+        "logo_path": null,
+        "name": "Darren Star Productions",
+        "origin_country": "US"
+      },
+      {
+        "id": 3268,
+        "logo_path": "/tuomPhY2UtuPTqqFnKMVHvSb724.png",
+        "name": "HBO",
+        "origin_country": "US"
+      }
+    ],
+    "production_countries": [
+      {
+        "iso_3166_1": "US",
+        "name": "United States of America"
+      }
+    ],
+    "seasons": [
+      {
+        "air_date": "2003-10-01",
+        "episode_count": 4,
+        "id": 105071,
+        "name": "Specials",
+        "overview": "",
+        "poster_path": "/s3JtxAtad3z2rzyDj2PAsGJ0KW3.jpg",
+        "season_number": 0,
+        "vote_average": 0
+      },
+      {
+        "air_date": "1998-06-06",
+        "episode_count": 12,
+        "id": 60249,
+        "name": "Season 1",
+        "overview": "Thirtysomething writer Carrie Bradshaw's experiences on the New York singles scene serve as the inspiration for her newspaper column: 'Sex and the City'. The struggles of finding a partner are seen through the eyes of Carrie and her best friends Samantha, Charlotte, and Miranda.",
+        "poster_path": "/cPxf5Y1aRE4iq5j08vd1y13cZMK.jpg",
+        "season_number": 1,
+        "vote_average": 5.6
+      },
+      {
+        "air_date": "1999-06-06",
+        "episode_count": 18,
+        "id": 60250,
+        "name": "Season 2",
+        "overview": "Carrie begins dating again in the wake of her break-up with Mr. Big while Samantha, Charlotte and Miranda continue to provide fodder for her column.",
+        "poster_path": "/1ZXJv019e5l2V0DWhD1IFGOGpKC.jpg",
+        "season_number": 2,
+        "vote_average": 6.3
+      },
+      {
+        "air_date": "2000-06-04",
+        "episode_count": 18,
+        "id": 60251,
+        "name": "Season 3",
+        "overview": "",
+        "poster_path": "/24JqcKtnduYEmsbTQo9upDM6cNz.jpg",
+        "season_number": 3,
+        "vote_average": 6.2
+      },
+      {
+        "air_date": "2001-06-03",
+        "episode_count": 18,
+        "id": 60252,
+        "name": "Season 4",
+        "overview": "",
+        "poster_path": "/ive3mic73CQqTGb5Hh9hLGNovfm.jpg",
+        "season_number": 4,
+        "vote_average": 6.1
+      },
+      {
+        "air_date": "2002-07-21",
+        "episode_count": 8,
+        "id": 60253,
+        "name": "Season 5",
+        "overview": "",
+        "poster_path": "/ixtl8ThPcDL7vd310uoHDdkt6vc.jpg",
+        "season_number": 5,
+        "vote_average": 6.4
+      },
+      {
+        "air_date": "2003-06-22",
+        "episode_count": 20,
+        "id": 60254,
+        "name": "Season 6",
+        "overview": "",
+        "poster_path": "/nYyY7AICJFCOhozkacJrHcYGbme.jpg",
+        "season_number": 6,
+        "vote_average": 5.7
+      }
+    ],
+    "spoken_languages": [
+      {
+        "english_name": "English",
+        "iso_639_1": "en",
+        "name": "English"
+      }
+    ],
+    "status": "Ended",
+    "tagline": "The next best thing to being there.",
+    "type": "Scripted",
+    "vote_average": 7.387,
+    "vote_count": 2070,
+    "keywords": {
+      "results": [
+        {
+          "name": "new york city",
+          "id": 242
+        },
+        {
+          "name": "beauty",
+          "id": 3258
+        },
+        {
+          "name": "female friendship",
+          "id": 5248
+        },
+        {
+          "name": "love",
+          "id": 9673
+        },
+        {
+          "name": "dating",
+          "id": 13043
+        },
+        {
+          "name": "newspaper reporter",
+          "id": 163098
+        },
+        {
+          "name": "manhattan, new york city",
+          "id": 168346
+        },
+        {
+          "name": "sitcom",
+          "id": 193171
+        }
+      ]
+    },
+    "similar": {
+      "page": 1,
+      "results": [
+        {
+          "adult": false,
+          "backdrop_path": null,
+          "genre_ids": [
+            35
+          ],
+          "id": 4263,
+          "origin_country": [
+            "US"
+          ],
+          "original_language": "en",
+          "original_name": "The New Dick Van Dyke Show",
+          "overview": "The New Dick Van Dyke Show is an American sitcom starring Dick Van Dyke that aired on CBS from 1971 to 1974. It was Van Dyke's first return to series television since The Dick Van Dyke Show.",
+          "popularity": 12.0966,
+          "poster_path": "/enC1neRtm1mGHnQtMIAvwhCnKO0.jpg",
+          "first_air_date": "1971-09-18",
+          "name": "The New Dick Van Dyke Show",
+          "vote_average": 4,
+          "vote_count": 3
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/azDlPDaWs0d7yxVA4Kr3ka8l2kE.jpg",
+          "genre_ids": [
+            35
+          ],
+          "id": 4277,
+          "origin_country": [
+            "US"
+          ],
+          "original_language": "en",
+          "original_name": "Archie Bunker's Place",
+          "overview": "Archie Bunker's Place is an American sitcom originally broadcast on the CBS network, conceived in 1979 as a spin-off and continuation of All in the Family. While not as popular as its predecessor, the show maintained a large enough audience to last for four seasons, until its cancellation in 1983. In its first season, the show performed so well that it knocked Mork & Mindy out of its new Sunday night time slot.",
+          "popularity": 22.9298,
+          "poster_path": "/kjNp6j9XIfUtNrIJWOC9uzfvZN4.jpg",
+          "first_air_date": "1979-09-23",
+          "name": "Archie Bunker's Place",
+          "vote_average": 6.5,
+          "vote_count": 20
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/hS2xAK3X1oWiaA8OX5Bow2SF6KN.jpg",
+          "genre_ids": [
+            35
+          ],
+          "id": 4292,
+          "origin_country": [
+            "GB"
+          ],
+          "original_language": "en",
+          "original_name": "Bottom",
+          "overview": "Richie Richard (socially awkward, sexually inexperienced) and Eddie Hitler (carefree alcoholic ) are two social outcasts living on the dole. Trapped together in a squalid flat in Hammersmith, London they are perpetually skint, bored and sexually frustrated. They spend their days scheming, bickering, and being nasty and sadistic to each other.",
+          "popularity": 7.5291,
+          "poster_path": "/bOrajiWj3Rax8KZ4ZZWcw2KpJD6.jpg",
+          "first_air_date": "1991-09-17",
+          "name": "Bottom",
+          "vote_average": 7.9,
+          "vote_count": 128
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/izWGnAJpD6Lthmq39s7Uyja2RgW.jpg",
+          "genre_ids": [
+            99,
+            10759,
+            80,
+            18
+          ],
+          "id": 4297,
+          "origin_country": [
+            "CA",
+            "US"
+          ],
+          "original_language": "en",
+          "original_name": "The Equalizer",
+          "overview": "Robert McCall is a former agent of a secret government agency who is now running his own private crime fighting operation where he fashions himself as \"The Equalizer.\" It is a service for victims of the system who have exhausted all possible means of seeking justice and have nowhere to go. McCall promises to even out the odds for them.",
+          "popularity": 33.2247,
+          "poster_path": "/7nrKPxGJWYe5cacNN9AV9bVplNZ.jpg",
+          "first_air_date": "1985-09-18",
+          "name": "The Equalizer",
+          "vote_average": 7.197,
+          "vote_count": 61
+        },
+        {
+          "adult": false,
+          "backdrop_path": "/jrOUuhskiGNiHbEcAQv8xsh4uPF.jpg",
+          "genre_ids": [
+            35
+          ],
+          "id": 4298,
+          "origin_country": [
+            "US"
+          ],
+          "original_language": "en",
+          "original_name": "Campus Ladies",
+          "overview": "Campus Ladies is an American sitcom that premiered on Oxygen on January 8, 2006. It stars Christen Sussin and Carrie Aizley as two 40-something women who decide to go enroll at the fictional University of the Midwest. The humor comes from the various situations the \"ladies\" get involved in while trying to fit in with their much younger friends and classmates.",
+          "popularity": 8.6864,
+          "poster_path": "/nxvJawrY41TIWT9hCLHsM56Syyf.jpg",
+          "first_air_date": "2006-01-08",
+          "name": "Campus Ladies",
+          "vote_average": 3.5,
+          "vote_count": 2
+        }
+      ],
+      "total_pages": 3278,
+      "total_results": 65543
+    }
+  }
+  """#
