--- Tests/TMDBTests/TVShows/Data+TrendingTVShows.swift
@@ -1,464 +1,0 @@
-import Foundation
-
-extension Data {
-  static let trendingTVShows = Data(content.utf8)
-}
-
-private let content = #"""
-  {
-    "page": 1,
-    "results": [
-      {
-        "adult": false,
-        "backdrop_path": "/cWO5NDkKqpOuwxu4vFc4PtL8aNF.jpg",
-        "id": 241388,
-        "name": "Eyes of Wakanda",
-        "original_name": "Eyes of Wakanda",
-        "overview": "Follow the adventures of brave Wakandan warriors throughout history in this globe-trotting adventure, where they must carry out dangerous missions to retrieve Vibranium artifacts from the enemies of Wakanda. They are the Hatut Zaraze and this is their story.",
-        "poster_path": "/yuOfb1MgnaGPa4guzV0n1IFYVGN.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          16,
-          10759,
-          10765
-        ],
-        "popularity": 73.02,
-        "first_air_date": "2025-08-01",
-        "vote_average": 4.583,
-        "vote_count": 24,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/2c3PCbxVgWen5HrYsaukUmfu4J9.jpg",
-        "id": 196890,
-        "name": "Chief of War",
-        "original_name": "Chief of War",
-        "overview": "With Hawaii's four kingdoms divided by war, the ferocious warrior Kaʻiana embarks on an epic mission to unite his people—as an existential threat approaches their shores.",
-        "poster_path": "/8Mckh3qJRTzXTQNZtpb133RHmO4.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          18
-        ],
-        "popularity": 103.9231,
-        "first_air_date": "2025-07-31",
-        "vote_average": 6.8,
-        "vote_count": 17,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/l5jzuehwynnoEWUb9z25auHXQM.jpg",
-        "id": 259909,
-        "name": "Dexter: Resurrection",
-        "original_name": "Dexter: Resurrection",
-        "overview": "Dexter Morgan awakens from a coma to find Harrison gone without a trace. Realizing the weight of what he put his son through, Dexter sets out for New York City, determined to find him and make things right. But closure won't come easy. When Miami Metro's Angel Batista arrives with questions, Dexter realizes his past is catching up to him fast. As father and son navigate their own darkness in the city that never sleeps, they soon find themselves deeper than they ever imagined - and that the only way out is together.",
-        "poster_path": "/hIawSocuwqgNeRf3JuKuxgMHmSC.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          80,
-          18
-        ],
-        "popularity": 88.0167,
-        "first_air_date": "2025-07-13",
-        "vote_average": 9.117,
-        "vote_count": 90,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/k6Yg53yBdsfZeRLJh0WgrG2P11f.jpg",
-        "id": 91097,
-        "name": "Ling Cage",
-        "original_name": "灵笼",
-        "overview": "In the near future, human inhabitants would have been crowded and congested. It was an urgency to stride out to the universe and find a new home. When everything was under progress in an orderly way, dramatic geological transformations erupted over the courses of decades. Humanity was demolished by this disaster and hardly left anything. Until the nature gradually restored calm, people struggled to their feet from ruins and abysses, stepping again onto this familiar but strange earth. But for us people, dominating everything has been rooted into our blood. Are we still masters of this new world?",
-        "poster_path": "/p65f4XPIneEjFhASI9HuJId3D3d.jpg",
-        "media_type": "tv",
-        "original_language": "zh",
-        "genre_ids": [
-          16,
-          10765,
-          10759,
-          18,
-          9648
-        ],
-        "popularity": 55.9796,
-        "first_air_date": "2019-07-13",
-        "vote_average": 8.297,
-        "vote_count": 64,
-        "origin_country": [
-          "CN"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/tOpyQG9MAnFN4YiDgyk5NnCGa7X.jpg",
-        "id": 243224,
-        "name": "The Immortal Ascension",
-        "original_name": "凡人修仙传",
-        "overview": "Han Li, an ordinary, poor boy from a mountain village, joins a small martial arts sect by chance. Despite his mediocre talent, he eventually cultivates himself into an immortal through hard work and careful planning.",
-        "poster_path": "/kG8YooBxqX5BByifbA0wYeVVRYe.jpg",
-        "media_type": "tv",
-        "original_language": "zh",
-        "genre_ids": [
-          18,
-          10765
-        ],
-        "popularity": 89.4534,
-        "first_air_date": "2025-07-27",
-        "vote_average": 6.667,
-        "vote_count": 12,
-        "origin_country": [
-          "CN"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/uxrP5yB9h6Dh6SbojI018pMYIIH.jpg",
-        "id": 90802,
-        "name": "The Sandman",
-        "original_name": "The Sandman",
-        "overview": "After years of imprisonment, Morpheus — the King of Dreams — embarks on a journey across worlds to find what was stolen from him and restore his power.",
-        "poster_path": "/vfJ1nBnqiiRWZWv3ZvtFWO5zccg.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          10765,
-          18,
-          10759
-        ],
-        "popularity": 116.0442,
-        "first_air_date": "2022-08-05",
-        "vote_average": 7.9,
-        "vote_count": 2301,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/xmDhagFdAfbkDFL6YpV3Vfe9OMl.jpg",
-        "id": 133748,
-        "name": "Twisted Metal",
-        "original_name": "Twisted Metal",
-        "overview": "A motor-mouthed outsider with no memory of his past is offered a chance at a better life, but only if he can successfully deliver a mysterious package across a post-apocalyptic wasteland. With the help of a badass axe-wielding car thief, he’ll face savage marauders driving vehicles of destruction and other dangers of the open road, including a deranged clown who drives an all too familiar ice cream truck.",
-        "poster_path": "/3Zr9qJXqHQwi9IyPBDwukgg1jcZ.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          10759,
-          35
-        ],
-        "popularity": 56.6641,
-        "first_air_date": "2023-07-27",
-        "vote_average": 7.855,
-        "vote_count": 400,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/9q7uryMCmAun2bGXoP3Liw24mpz.jpg",
-        "id": 296834,
-        "name": "Nine Yin True Sutra",
-        "original_name": "华山论剑",
-        "overview": "",
-        "poster_path": "/gwHpqRxpe0o0pDmDZK7V8RVMhTJ.jpg",
-        "media_type": "tv",
-        "original_language": "zh",
-        "genre_ids": [
-          18
-        ],
-        "popularity": 15.9526,
-        "first_air_date": "2025-07-26",
-        "vote_average": 5,
-        "vote_count": 4,
-        "origin_country": [
-          "CN"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/oUisxaYRPuIehH6hkUIMcZNIuTA.jpg",
-        "id": 93740,
-        "name": "Foundation",
-        "original_name": "Foundation",
-        "overview": "Follow a band of exiles on their monumental journey to save humanity and rebuild civilization amid the fall of the Galactic Empire.",
-        "poster_path": "/tg9I5pOY4M9CKj8U0cxVBTsm5eh.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          10765,
-          18
-        ],
-        "popularity": 94.1003,
-        "first_air_date": "2021-09-23",
-        "vote_average": 7.8,
-        "vote_count": 1431,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/kWByUrGgS4mYyvQ2mG95gzblBOj.jpg",
-        "id": 106449,
-        "name": "A Record of a Mortal's Journey to Immortality",
-        "original_name": "凡人修仙传",
-        "overview": "A poor and ordinary boy from a village joins a minor sect in Jiang Hu and becomes an Unofficial Disciple by chance. How will Han Li, a commoner by birth, establish a foothold for himself in his sect? With his mediocre aptitude, he must successfully traverse the treacherous path of cultivation and avoid the notice of those who may do him harm. This is a story of an ordinary mortal who, against all odds, clashes with devilish demons and ancient celestials in order to find his own path towards immortality.",
-        "poster_path": "/3m1UaMLgmpj6krNaQwTpftYFsnz.jpg",
-        "media_type": "tv",
-        "original_language": "zh",
-        "genre_ids": [
-          16,
-          10759,
-          10765
-        ],
-        "popularity": 37.6048,
-        "first_air_date": "2020-07-25",
-        "vote_average": 9.2,
-        "vote_count": 45,
-        "origin_country": [
-          "CN"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/bqtIpYB9sfWJ0jGmSSGXiATD4ew.jpg",
-        "id": 253955,
-        "name": "The Wanted Detective",
-        "original_name": "定风波",
-        "overview": "Framed as the notorious Night Devil on his wedding day, top detective Xiao Beiming is accused of murder and treason, forcing him to flee. Three years later, he returns to the capital, earning the trust of his apprentice sister, Zhong Xueman. With a team of fearless allies, he unravels a series of mysterious cases, to expose the real Night Devil and restore order to Da Qi.",
-        "poster_path": "/jd105sxPMO3KfuMUI1RecTeG6K5.jpg",
-        "media_type": "tv",
-        "original_language": "zh",
-        "genre_ids": [
-          18,
-          9648
-        ],
-        "popularity": 7.4305,
-        "first_air_date": "2025-08-01",
-        "vote_average": 0,
-        "vote_count": 0,
-        "origin_country": [
-          "CN"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/2rmK7mnchw9Xr3XdiTFSxTTLXqv.jpg",
-        "id": 37854,
-        "name": "One Piece",
-        "original_name": "ワンピース",
-        "overview": "Years ago, the fearsome Pirate King, Gol D. Roger was executed leaving a huge pile of treasure and the famous \"One Piece\" behind. Whoever claims the \"One Piece\" will be named the new King of the Pirates.\n\nMonkey D. Luffy, a boy who consumed a \"Devil Fruit,\" decides to follow in the footsteps of his idol, the pirate Shanks, and find the One Piece. It helps, of course, that his body has the properties of rubber and that he's surrounded by a bevy of skilled fighters and thieves to help him along the way.\n\nLuffy will do anything to get the One Piece and become King of the Pirates!",
-        "poster_path": "/cMD9Ygz11zjJzAovURpO75Qg7rT.jpg",
-        "media_type": "tv",
-        "original_language": "ja",
-        "genre_ids": [
-          10759,
-          35,
-          16
-        ],
-        "popularity": 71.377,
-        "first_air_date": "1999-10-20",
-        "vote_average": 8.724,
-        "vote_count": 4951,
-        "origin_country": [
-          "JP"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/uNTrRKIOyKYISthoeizghtXPEOK.jpg",
-        "id": 240411,
-        "name": "Dan Da Dan",
-        "original_name": "ダンダダン",
-        "overview": "In a bet to prove whether ghosts or aliens exist, two high schoolers face terrifying paranormal threats, gain superpowers and maybe even fall in love?!",
-        "poster_path": "/6qfZAOEUFIrbUH3JvePclx1nXzz.jpg",
-        "media_type": "tv",
-        "original_language": "ja",
-        "genre_ids": [
-          16,
-          10759,
-          35,
-          10765
-        ],
-        "popularity": 122.195,
-        "first_air_date": "2024-10-04",
-        "vote_average": 8.643,
-        "vote_count": 582,
-        "origin_country": [
-          "JP"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/maxJ6oaC3tgO0lUiudLCLN4FPS8.jpg",
-        "id": 293808,
-        "name": "Marked",
-        "original_name": "Marked",
-        "overview": "Desperate to pay for her daughter's life-saving surgery, devout ex-cop Babalwa turns to an unholy savior to help her stage a daring heist.",
-        "poster_path": "/1XFP7vl4KDAuCNOdtq8lFm5LeSW.jpg",
-        "media_type": "tv",
-        "original_language": "zu",
-        "genre_ids": [
-          18
-        ],
-        "popularity": 18.2129,
-        "first_air_date": "2025-07-31",
-        "vote_average": 5.5,
-        "vote_count": 4,
-        "origin_country": [
-          "ZA"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/6omimfVty4JD3LXzI5LHEGhXjwn.jpg",
-        "id": 284445,
-        "name": "Takopi's Original Sin",
-        "original_name": "タコピーの原罪",
-        "overview": "A Happy alien, Takopi, lands on Earth with one mission: to spread happiness! When he meets Shizuka, a lonely fourth grader, he vows to bring back her smile using his magical Happy Gadgets. But as he uncovers the pain in her life, Takopi learns that true happiness may require more than gadgets.",
-        "poster_path": "/xPXDVhVKt0XM34ihoUVMHtLYTw8.jpg",
-        "media_type": "tv",
-        "original_language": "ja",
-        "genre_ids": [
-          16,
-          18,
-          10765
-        ],
-        "popularity": 36.9203,
-        "first_air_date": "2025-06-28",
-        "vote_average": 9.1,
-        "vote_count": 41,
-        "origin_country": [
-          "JP"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/6WAinokBmZU9Jt8l74TKI8JNUG7.jpg",
-        "id": 2190,
-        "name": "South Park",
-        "original_name": "South Park",
-        "overview": "Follow the misadventures of four irreverent grade-schoolers in the quiet, dysfunctional town of South Park, Colorado.",
-        "poster_path": "/dqAKFEjPEu7vWkVBkMdhGu9ifxU.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          16,
-          35
-        ],
-        "popularity": 146.0586,
-        "first_air_date": "1997-08-13",
-        "vote_average": 8.358,
-        "vote_count": 4610,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/2meX1nMdScFOoV4370rqHWKmXhY.jpg",
-        "id": 93405,
-        "name": "Squid Game",
-        "original_name": "오징어 게임",
-        "overview": "Hundreds of cash-strapped players accept a strange invitation to compete in children's games. Inside, a tempting prize awaits — with deadly high stakes.",
-        "poster_path": "/1QdXdRYfktUSONkl1oD5gc6Be0s.jpg",
-        "media_type": "tv",
-        "original_language": "ko",
-        "genre_ids": [
-          10759,
-          9648,
-          18
-        ],
-        "popularity": 301.2045,
-        "first_air_date": "2021-09-17",
-        "vote_average": 7.865,
-        "vote_count": 16491,
-        "origin_country": [
-          "KR"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/i5qNKvdBkVAJvJZRkkzSvQxrAL2.jpg",
-        "id": 252751,
-        "name": "Leanne",
-        "original_name": "Leanne",
-        "overview": "When her husband leaves her for another woman, a resilient Southern mom starts over with the support of her loving, but unfiltered family.",
-        "poster_path": "/sILKrYhFK6bedXgDggQD3GmGUUV.jpg",
-        "media_type": "tv",
-        "original_language": "en",
-        "genre_ids": [
-          35
-        ],
-        "popularity": 20.9798,
-        "first_air_date": "2025-07-31",
-        "vote_average": 10,
-        "vote_count": 4,
-        "origin_country": [
-          "US"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/9DrORnlwnjt5SeJKcuE5sqOoIdV.jpg",
-        "id": 207468,
-        "name": "Kaiju No. 8",
-        "original_name": "怪獣８号",
-        "overview": "In a world plagued by creatures known as Kaiju, Kafka Hibino aspired to enlist in The Defense Force. He makes a promise to enlist with his childhood friend, Mina Ashiro. Soon, life takes them in separate ways. While employed cleaning up after Kaiju battles, Kafka meets Reno Ichikawa. Reno's determination to join The Defense Force reawakens Kafka's promise to join Mina and protect humanity.",
-        "poster_path": "/PDe8WZHxH6XxM6Buxoo99aQrgw.jpg",
-        "media_type": "tv",
-        "original_language": "ja",
-        "genre_ids": [
-          16,
-          10759,
-          10765
-        ],
-        "popularity": 94.7975,
-        "first_air_date": "2024-04-13",
-        "vote_average": 8.561,
-        "vote_count": 578,
-        "origin_country": [
-          "JP"
-        ]
-      },
-      {
-        "adult": false,
-        "backdrop_path": "/y6re1YeUJJIAOhCG7CzFbnX70Mq.jpg",
-        "id": 1429,
-        "name": "Attack on Titan",
-        "original_name": "進撃の巨人",
-        "overview": "Many years ago, the last remnants of humanity were forced to retreat behind the towering walls of a fortified city to escape the massive, man-eating Titans that roamed the land outside their fortress. Only the heroic members of the Scouting Legion dared to stray beyond the safety of the walls – but even those brave warriors seldom returned alive. Those within the city clung to the illusion of a peaceful existence until the day that dream was shattered, and their slim chance at survival was reduced to one horrifying choice: kill – or be devoured!",
-        "poster_path": "/hTP1DtLGFamjfu8WqjnuQdP1n4i.jpg",
-        "media_type": "tv",
-        "original_language": "ja",
-        "genre_ids": [
-          16,
-          10765,
-          10759
-        ],
-        "popularity": 42.9566,
-        "first_air_date": "2013-04-07",
-        "vote_average": 8.669,
-        "vote_count": 6882,
-        "origin_country": [
-          "JP"
-        ]
-      }
-    ],
-    "total_pages": 500,
-    "total_results": 10000
-  }
-  """#
