+++ Tests/TMDBTests/TVShows/TVShowClientTests.swift
@@ -1,0 +1,76 @@
+//
+//  TVShowClientTests.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+import CustomDump
+import Dependencies
+import Foundation
+import Testing
+
+@testable import TMDB
+
+#if canImport(FoundationNetworking)
+  import FoundationNetworking
+#endif
+
+struct TVShowClientTests {
+  @Test func register() async throws {
+    let detailsCalls = LockIsolated<[(TVShow.ID, [TVShowAppendOptions], Locale?, AccessToken)]>([])
+    _ = try await withDependencies {
+      $0.tvShowClient.details = { id, appending, locale, accessToken in
+        detailsCalls.withValue { $0.append((id, appending, locale, accessToken)) }
+        return .mock
+      }
+    } operation: {
+      @Dependency(\.tvShowClient) var tvShowClient
+      return try await tvShowClient.details(456, [.keywords], Locale(identifier: "en"), "DEF456")
+    }
+    #expect(detailsCalls.value.count == 1)
+    let call = try #require(detailsCalls.value.first)
+    expectNoDifference(call.0, 456)
+    expectNoDifference(call.1, [.keywords])
+    expectNoDifference(call.2, Locale(identifier: "en"))
+    expectNoDifference(call.3, "DEF456")
+  }
+
+  @Test func detailsSuccess() async throws {
+    let fetchCalls = LockIsolated<[URLRequest]>([])
+    let sharedClient = SharedClient(
+      fetch: { request in
+        fetchCalls.withValue { $0.append(request) }
+        let response = try HTTPURLResponse(
+          url: #require(URL(string: "https://api.themoviedb.org/3/tv/456")),
+          statusCode: 200,
+          httpVersion: nil,
+          headerFields: nil
+        )
+        return try (.tvShowDetailsComplete, #require(response))
+      }
+    )
+    let tvShowClient = TVShowClient.liveValue
+    let details = try await withDependencies {
+      $0.sharedClient = sharedClient
+    } operation: {
+      try await tvShowClient.details(456, [.keywords], Locale(identifier: "en"), "DEF456")
+    }
+    expectNoDifference(details, .mock)
+    #expect(fetchCalls.value.count == 1)
+    let urlRequest = try #require(fetchCalls.value.first)
+    expectNoDifference(urlRequest, expectedURLRequest())
+  }
+}
+
+private func expectedURLRequest() -> URLRequest? {
+  let url = URL(
+    string:
+      "https://api.themoviedb.org/3/tv/ID(rawValue:%20456)?append_to_response=keywords&language=en")!
+  var urlRequest = URLRequest(url: url)
+  urlRequest.allHTTPHeaderFields = [
+    "Accept": "application/json",
+    "Authorization": "Bearer DEF456",
+  ]
+  return urlRequest
+}
