--- Tests/TMDBTests/TestStorage.swift
@@ -1,7 +1,0 @@
-actor TestStorage<T> {
-  var value: T?
-
-  func setValue(_ value: T) {
-    self.value = value
-  }
-}
